Generate list of field names subset by parameter

Function subsets the input dataframe by the input parameter and creates a list of the fields in the subset dataframe as well as the number of unique values in each field. The list is intended to inform users as to which specific fields to explore further and filter subset by a parameter.

Usage

FilterParFields(.data, parameter)

Arguments

.data

TADA dataframe

parameter

Characteristic name (parameter name) from the dataframe.

Value

A table of fields and the count of unique values in each field, subset by a parameter.

Examples

# Load example dataset:
data(Nutrients_Utah)

# Create list of fields for parameter "AMMONIA" with number of unique values in each field:
AmmoniaFields <- FilterParFields(Nutrients_Utah, parameter = "AMMONIA")
#>                                                                Fields Count
#> 1                                              OrganizationIdentifier     2
#> 2                                              OrganizationFormalName     2
#> 3                                                  ActivityIdentifier    68
#> 4                                                    ActivityTypeCode     5
#> 5                                                   ActivityMediaName     1
#> 6                                        ActivityMediaSubdivisionName     1
#> 7                                                   ActivityStartDate    29
#> 8                                              ActivityStartTime.Time    56
#> 9                                      ActivityStartTime.TimeZoneCode     1
#> 10                                                    ActivityEndDate     1
#> 11                                               ActivityEndTime.Time     1
#> 12                                       ActivityEndTime.TimeZoneCode     1
#> 13                                          ActivityRelativeDepthName     1
#> 14                            ActivityDepthHeightMeasure.MeasureValue     1
#> 15                         ActivityDepthHeightMeasure.MeasureUnitCode     1
#> 16                            ActivityDepthAltitudeReferencePointText     1
#> 17                         ActivityTopDepthHeightMeasure.MeasureValue     1
#> 18                      ActivityTopDepthHeightMeasure.MeasureUnitCode     1
#> 19                      ActivityBottomDepthHeightMeasure.MeasureValue     1
#> 20                   ActivityBottomDepthHeightMeasure.MeasureUnitCode     1
#> 21                                                  ProjectIdentifier     3
#> 22                                                        ProjectName     3
#> 23                                 ActivityConductingOrganizationText     1
#> 24                                       MonitoringLocationIdentifier    27
#> 25                                             MonitoringLocationName    21
#> 26                                                ActivityCommentText     1
#> 27                                                      SampleAquifer     1
#> 28                                                HydrologicCondition     1
#> 29                                                    HydrologicEvent     1
#> 30                                   ActivityLocation.LatitudeMeasure    27
#> 31                                  ActivityLocation.LongitudeMeasure    27
#> 32                            SampleCollectionMethod.MethodIdentifier     2
#> 33                     SampleCollectionMethod.MethodIdentifierContext     2
#> 34                                  SampleCollectionMethod.MethodName     2
#> 35                       SampleCollectionMethod.MethodDescriptionText     2
#> 36                                      SampleCollectionEquipmentName     1
#> 37                                                   ResultIdentifier    72
#> 38                                       ResultDetectionConditionText     3
#> 39                                               MethodSpeciationName     2
#> 40                                                 CharacteristicName     1
#> 41                                           ResultSampleFractionText     2
#> 42                                                 ResultMeasureValue     9
#> 43                                      ResultMeasure.MeasureUnitCode     1
#> 44                                               MeasureQualifierCode     3
#> 45                                             ResultStatusIdentifier     1
#> 46                                                StatisticalBaseCode     1
#> 47                                                ResultValueTypeName     1
#> 48                                              ResultWeightBasisText     1
#> 49                                                ResultTimeBasisText     1
#> 50                                         ResultTemperatureBasisText     1
#> 51                                        ResultParticleSizeBasisText     1
#> 52                                         DataQuality.PrecisionValue     1
#> 53                                              DataQuality.BiasValue     1
#> 54                                DataQuality.ConfidenceIntervalValue     1
#> 55                              DataQuality.UpperConfidenceLimitValue     1
#> 56                              DataQuality.LowerConfidenceLimitValue     1
#> 57                                                  ResultCommentText     1
#> 58                                                          USGSPCode     1
#> 59                              ResultDepthHeightMeasure.MeasureValue     1
#> 60                           ResultDepthHeightMeasure.MeasureUnitCode     1
#> 61                              ResultDepthAltitudeReferencePointText     1
#> 62                                               SubjectTaxonomicName     1
#> 63                                            SampleTissueAnatomyName     1
#> 64                                               BinaryObjectFileName     1
#> 65                                           BinaryObjectFileTypeCode     1
#> 66                                                      ResultFileUrl     1
#> 67                            ResultAnalyticalMethod.MethodIdentifier     1
#> 68                     ResultAnalyticalMethod.MethodIdentifierContext     1
#> 69                                  ResultAnalyticalMethod.MethodName     1
#> 70                                   ResultAnalyticalMethod.MethodUrl     1
#> 71                       ResultAnalyticalMethod.MethodDescriptionText     1
#> 72                                                     LaboratoryName     3
#> 73                                                  AnalysisStartDate    24
#> 74                                        ResultLaboratoryCommentText     1
#> 75                                ResultDetectionQuantitationLimitUrl    72
#> 76                                 DetectionQuantitationLimitTypeName     3
#> 77                     DetectionQuantitationLimitMeasure.MeasureValue     3
#> 78                  DetectionQuantitationLimitMeasure.MeasureUnitCode     1
#> 79                                            LabSamplePreparationUrl     1
#> 80                                                        LastUpdated    48
#> 81                                                       ProviderName     1
#> 82                                                      timeZoneStart     1
#> 83                                                        timeZoneEnd     1
#> 84                                              ActivityStartDateTime    62
#> 85                                                ActivityEndDateTime     1
#> 86                                         MonitoringLocationTypeName     3
#> 87                                  MonitoringLocationDescriptionText    22
#> 88                                                  HUCEightDigitCode     5
#> 89                                   DrainageAreaMeasure.MeasureValue     1
#> 90                                DrainageAreaMeasure.MeasureUnitCode     1
#> 91                       ContributingDrainageAreaMeasure.MeasureValue     1
#> 92                    ContributingDrainageAreaMeasure.MeasureUnitCode     1
#> 93                                                    LatitudeMeasure    27
#> 94                                                   LongitudeMeasure    27
#> 95                                              SourceMapScaleNumeric     1
#> 96                             HorizontalAccuracyMeasure.MeasureValue     1
#> 97                          HorizontalAccuracyMeasure.MeasureUnitCode     1
#> 98                                     HorizontalCollectionMethodName     5
#> 99                       HorizontalCoordinateReferenceSystemDatumName     4
#> 100                                      VerticalMeasure.MeasureValue     8
#> 101                                   VerticalMeasure.MeasureUnitCode     2
#> 102                              VerticalAccuracyMeasure.MeasureValue     1
#> 103                           VerticalAccuracyMeasure.MeasureUnitCode     1
#> 104                                      VerticalCollectionMethodName     3
#> 105                        VerticalCoordinateReferenceSystemDatumName     2
#> 106                                                       CountryCode     1
#> 107                                                         StateCode     1
#> 108                                                        CountyCode     3
#> 109                                                       AquiferName     1
#> 110                                                     LocalAqfrName     1
#> 111                                                 FormationTypeText     1
#> 112                                                   AquiferTypeName     1
#> 113                                              ConstructionDateText     1
#> 114                                     WellDepthMeasure.MeasureValue     1
#> 115                                  WellDepthMeasure.MeasureUnitCode     1
#> 116                                 WellHoleDepthMeasure.MeasureValue     3
#> 117                              WellHoleDepthMeasure.MeasureUnitCode     2
#> 118                                           MethodSpecificationName     2
#> 119                                            ProjectDescriptionText     3
#> 120                                            SamplingDesignTypeCode     1
#> 121                                             QAPPApprovedIndicator     2
#> 122                                            QAPPApprovalAgencyName     1
#> 123                                                    ProjectFileUrl     1
#> 124                             ProjectMonitoringLocationWeightingUrl     1
#> 125                                              TADA.LatitudeMeasure    27
#> 126                                             TADA.LongitudeMeasure    27
#> 127                                            TADA.ActivityMediaName     1
#> 128                                           TADA.CharacteristicName     1
#> 129                                           TADA.ResultMeasureValue    11
#> 130                             TADA.ResultMeasureValueDataTypes.Flag     2
#> 131                                TADA.ResultMeasure.MeasureUnitCode     1
#> 132                                      TADA.MethodSpecificationName     2
#> 133                                     TADA.ResultSampleFractionText     2
#> 134               TADA.DetectionQuantitationLimitMeasure.MeasureValue     3
#> 135            TADA.DetectionQuantitationLimitMeasure.MeasureUnitCode     1
#> 136 TADA.DetectionQuantitationLimitMeasure.MeasureValueDataTypes.Flag     1