Transform CharacteristicName, ResultSampleFractionText, MethodSpecificationName, and ResultMeasure.MeasureUnitCode values to TADA standards.

Function compares input dataframe to the TADA Harmonization Reference Table, and makes synonymous data consistent. Optional outputs include: 1) the dataframe with Harmonization columns appended, 2) the dataframe with CharacteristicName, ResultSampleFractionText, MethodSpecificationName, and ResultMeasure.MeasureUnitCode converted to TADA standards or 3) the four fields converted with most Harmonization Reference Table columns appended. Default is transform = TRUE and flag = TRUE.

HarmonizeData(.data, ref, transform = TRUE, flag = TRUE)

Arguments

.data

TADA dataframe

ref

Optional argument to specify which dataframe to use as a reference file. The primary use for this argument is when a user has generated a harmonization reference file unique to their data, and they made changes to that file.

transform

Boolean argument; transforms and/or converts original values in the dataframe to the TADA Harmonization Reference Table values for the following fields: CharacteristicName, ResultSampleFractionText, MethodSpecificationName, and ResultMeasure.MeasureUnitCode. Default is transform = TRUE.

flag

Boolean argument; appends all columns from the TADA Harmonization Reference Table to the dataframe. Default is flag = TRUE.

Value

When transform = FALSE and flag = TRUE, Harmonization Reference Table columns are appended to the dataframe only. When transform = TRUE and flag = TRUE, Harmonization columns are appended to the dataframe and transformations are executed. When transform = TRUE and flag = FALSE, transformations are executed only. When transform = FALSE and flag = FALSE, an error is returned (function would return the input dataframe unchanged if input was allowed).