Invalid coordinates

This function identifies and flags invalid coordinate data. When clean_outsideUSA = FALSE and clean_imprecise = FALSE, a column will be appended titled "TADA.InvalidCoordinates" with the following flags: 1) If the latitude is less than zero, the row will be flagged with "LAT_OutsideUSA", 2) If the longitude is greater than zero AND less than 145, the row will be flagged as "LONG_OutsideUSA", 3) If the latitude or longitude contains the string, "999", the row will be flagged as invalid, and 4) Finally, precision can be measured by the number of decimal places in the latitude and longitude provided. If either the lattitue or longitude does not have any numbers to the right of the decimal point, the row will be flagged as "Imprecise".

InvalidCoordinates(.data, clean_outsideUSA = FALSE, clean_imprecise = FALSE)

Arguments

.data

TADA dataframe

clean_outsideUSA

Boolean argument; removes data with coordinates outside of the United States when clean_outsideUSA = TRUE. Default is clean = FALSE.

clean_imprecise

Boolean arguments; removes imprecise data when clean_imprecise = TRUE. Default is clean_imprecise = FALSE.

Value

When either the clean_outsideUSA or clean_imprecise argument is FALSE, a column flagging rows with the respective QA check is appended to the input dataframe. When either argument is TRUE, "invalid" or "imprecise" data is removed, respectively.