Check for Invalid Analytical Methods

Function checks the validity of each characteristic-analytical method combination in the dataframe. When clean = TRUE, rows with invalid characteristic-analytical method combinations are removed. Default is clean = TRUE. When errorsonly = TRUE, dataframe is filtered to show only invalid characteristic-analytical method combinations. Default is errorsonly = FALSE.

Usage

InvalidMethod(.data, clean = TRUE, errorsonly = FALSE)

Arguments

.data

TADA dataframe

clean

Boolean argument; removes "Invalid" characteristic-analytical method combinations from the dataframe when clean = TRUE. Default is clean = TRUE.

errorsonly

Boolean argument; filters dataframe to show only "Invalid" characteristic-analytical method combinations when errorsonly = TRUE. Default is errorsonly = FALSE.

Value

When clean = FALSE and errorsonly = FALSE, this function adds the following column to your dataframe: TADA.AnalyticalMethod.Flag. This column flags invalid CharacteristicName, ResultAnalyticalMethod/MethodIdentifier, and ResultAnalyticalMethod/MethodIdentifierContext combinations in your dataframe as either "Nonstandardized", "Invalid", or "Valid". When clean = FALSE and errorsonly = TRUE, the dataframe is filtered to show only "Invalid" characteristic-analytical method combinations; the column TADA.AnalyticalMethod.Flag is still appended. When clean = TRUE and errorsonly = FALSE, "Invalid" rows are removed from the dataframe and no column will be appended.

Examples

# Load example dataset
data(Nutrients_Utah)

# Remove invalid characteristic-analytical method combinations from dataframe:
InvalidMethod_clean <- InvalidMethod(Nutrients_Utah)
#> [1] "No invalid method/characteristic combinations in your dataframe. Returning the input dataframe with TADA.AnalyticalMethod.Flag column for tracking."

# Flag, but do not remove, invalid characteristic-analytical method combinations
# in new column titled "TADA.AnalyticalMethod.Flag":
InvalidMethod_flags <- InvalidMethod(Nutrients_Utah, clean = FALSE)
#> [1] "No invalid method/characteristic combinations in your dataframe. Returning the input dataframe with TADA.AnalyticalMethod.Flag column for tracking."

# Show only invalid characteristic-analytical method combinations:
InvalidMethod_errorsonly <- InvalidMethod(Nutrients_Utah, clean = FALSE, errorsonly = TRUE)
#> [1] "This dataframe is empty because we did not find any invalid method/characteristic combinations in your dataframe"