Large WQP data pulls using dataRetrieval

This function does multiple synchronous data calls to the WQP (waterqualitydata.us). It uses the WQP summary service to limit the amount downloaded to only relevant data, and pulls back data from 100 stations at a time and then joins the data back together and produces a single TADA compatible dataframe as the output. For large data sets, that can save a lot of time and ultimately reduce the complexity of subsequent data processing. Using this function, you will be able to download all data available from all sites in the contiguous United States that is available for the time period, characteristicName, and siteType requested.

TADABigdataRetrieval(
  startDate = "null",
  endDate = "null",
  characteristicName = "null",
  siteType = "null"
)

Arguments

startDate

Start Date YYYY-MM-DD format, for example, "1995-01-01"

endDate

end date in YYYY-MM-DD format, for example, "2020-12-31"

characteristicName

Name of water quality parameter

siteType

Name of water body type (e.g., "Stream", "Lake, Reservoir, Impoundment")

Value

TADA-compatible dataframe

Details

Similarly to the TADAdataRetrieval function, this function will create and/or edit the following columns: TADA.DetectionLimitMeasureValue.Flag DetectionQuantitationLimitMeasure.MeasureValue DetectionLimitMeasureValue.Original ResultMeasureValue.Original TADA.ResultMeasureValue.Flag ResultMeasureValue

All data cleaning and transformations are done directly to the "ResultMeasureValue" and "DetectionLimitMeasureValue" columns, however the original "ResultMeasureValue" and "DetectionLimitMeasureValue" columns and values from the WQP are preserved in these new fields, "ResultMeasureValue.Original" and "DetectionLimitMeasureValue.Original". Additionally, "TADA.ResultMeasureValue.Flag" and "TADA.DetectionLimitMeasureValue.Flag" are created to track and changes made to the "ResultMeasureValue" and "DetectionLimitMeasureValue" columns; and to provide information about the result values that is needed to address censored data later on (i.e., nondetections)

See ?MeasureValueSpecialCharacters and ?autoclean documentation for more information.

Examples

if (FALSE) {
tada2 <- TADABigdataRetrieval(startDate = "2019-01-01", endDate = "2021-12-31", characteristicName = "Temperature, water", siteType = "Stream")
}