Create Overview Map

Create Overview Map

Usage

TADAOverviewMap(.data)

Arguments

.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'MonitoringLocationIdentifier','MonitoringLocationName','TADA.LatitudeMeasure', 'TADA.LongitudeMeasure', 'ResultIdentifier', 'ActivityStartDate', 'TADA.CharacteristicName', and 'OrganizationIdentifier' to run this function.

Value

A leaflet map that shows all sites in the data frame, where larger point sizes indicate more results collected at a site, and darker point colors indicate more characteristics measured at that site. Users can click on points on the map to see a pop-up window with exact counts for results, characteristics, and organizations associated with each site.

Examples

# Load example dataset:
data("Nutrients_Utah.rda")
#> Warning: data set 'Nutrients_Utah.rda' not found

# Create map:
TADAOverviewMap(Nutrients_Utah)