Create Histogram(s)

Create Histogram(s)

Usage

TADA_Histogram(.data, id_cols = c("TADA.ComparableDataIdentifier"))

Arguments

.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'TADA.ComparableDataIdentifier', 'TADA.ResultMeasureValue', and 'TADA.ResultMeasure.MeasureUnitCode' to run this function. 'TADA.ComparableDataIdentifier' can be added to the data frame by running the function HarmonizeData(). The user can include additional grouping columns in the id_col input. If more than one group exists in the dataset (i.e. two or more unique comparable data identifiers), the function creates a list of plots, where each list element name is a unique group identifier.

id_cols

The column(S) in the dataset used to identify the unique groups to be plotted. Defaults to 'TADA.ComparableDataIdentifier'.

Value

A list of plotly histogram figures showing the distribution of sample values for each data group.

Examples

# Load example dataset:
data(Data_6Tribes_5y_Harmonized)

# Create a histogram for each comparable data group (TADA.ComparableDataIdentifier)
# in the input dataframe:
TADA_Histogram(Data_6Tribes_5y_Harmonized, id_cols = "TADA.ComparableDataIdentifier")
#> $`ALKALINITY, TOTAL_NA_NA_UG/L`
#> 
#> $`ALKALINITY, TOTAL_TOTAL_AS CACO3_PCU`
#> 
#> $`ALKALINITY, TOTAL_TOTAL_AS CACO3_UG/L`
#> 
#> $`ALKALINITY, TOTAL_TOTAL_NA_UG/L`
#> 
#> $`ALPHA PARTICLE_DISSOLVED_NA_PCI/L`
#> 
#> $`ALPHA PARTICLE_NA_NA_PCI/L`
#> 
#> $`ALPHA PARTICLE_TOTAL_NA_PCI/L`
#> 
#> $`ALUMINUM_DISSOLVED_NA_UG/L`
#> 
#> $`AMMONIA-NITROGEN_DISSOLVED_AS NH3_UG/L`
#> 
#> $`AMMONIA-NITROGEN_DISSOLVED_AS N_UG/L`
#> 
#> $`AMMONIA-NITROGEN_TOTAL_AS NH3_UG/L`
#> 
#> $`AMMONIA-NITROGEN_TOTAL_AS N_UG/L`
#> 
#> $`AMMONIA-NITROGEN_TOTAL_NA_UG/L`
#> 
#> $`AMMONIA-NITROGEN_TOTAL_NONE_UG/L`
#> 
#> $`AMMONIA_TOTAL_NA_UG/L`
#> 
#> $`AMMONIUM_TOTAL_NA_UG/L`
#> 
#> $`APPARENT COLOR_TOTAL_NA_PCU`
#> 
#> $`ARSENIC_DISSOLVED_NA_UG/L`
#> 
#> $`ARSENIC_TOTAL_NA_UG/L`
#> 
#> $`BARIUM_DISSOLVED_NA_UG/L`
#> 
#> $`BAROMETRIC PRESSURE_NA_NA_G/M2`
#> 
#> $`BENZENE HEXACHLORIDE, ALPHA (BHC)_TOTAL_NA_UG/L`
#> 
#> $`BERYLLIUM_DISSOLVED_NA_UG/L`
#> 
#> $`BHC, .BETA.-BHC & .GAMMA.-BHC MIX, UNSPECIFIED_TOTAL_NA_UG/L`
#> 
#> $`BIOCHEMICAL OXYGEN DEMAND, STANDARD CONDITIONS_TOTAL_NA_UG/L`
#> 
#> $`BORON_DISSOLVED_NA_UG/L`
#> 
#> $`CADMIUM_DISSOLVED_NA_UG/L`
#> 
#> $`CADMIUM_TOTAL_NA_UG/L`
#> 
#> $`CALCIUM_DISSOLVED_NA_UG/L`
#> 
#> $`CHEMICAL OXYGEN DEMAND_TOTAL_NA_UG/L`
#> 
#> $`CHLORIDE_TOTAL_NA_UG/L`
#> 
#> $`CHLOROPHYLL A, CORRECTED FOR PHEOPHYTIN_SUSPENDED_NA_UG/L`
#> 
#> $`CHLOROPHYLL A_NA_NA_UG/L`
#> 
#> $`CHLOROPHYLL A_TOTAL_NA_UG/L`
#> 
#> $`CHROMIUM_DISSOLVED_NA_UG/L`
#> 
#> $`CHROMIUM_TOTAL_NA_UG/L`
#> 
#> $`COBALT_DISSOLVED_NA_UG/L`
#> 
#> $`CONDUCTIVITY_NA_NA_US/CM`
#> 
#> $`COPPER_DISSOLVED_NA_UG/L`
#> 
#> $`COPPER_TOTAL_NA_UG/L`
#> 
#> $COUNT_NA_NA_COUNT
#> 
#> $`CYANIDE_DISSOLVED_NA_UG/L`
#> 
#> $`CYANIDE_TOTAL_NA_UG/L`
#> 
#> $`DEPTH, SECCHI DISK DEPTH_NA_NA_FT`
#> 
#> $`DEPTH, SECCHI DISK DEPTH_NA_NA_M`
#> 
#> $`DEPTH, SNOW COVER_NA_NA_IN`
#> 
#> $DEPTH_NA_NA_FT
#> 
#> $DEPTH_NA_NA_IN
#> 
#> $DEPTH_NA_NA_M
#> 
#> $`DISCHARGE, RIVER/STREAM_NA_NA_CFS`
#> 
#> $`DISSOLVED OXYGEN (DO)_NA_NA_UG/L`
#> 
#> $`DISSOLVED OXYGEN SATURATION_NA_NA_%`
#> 
#> $`DISSOLVED OXYGEN UPTAKE_NA_NA_UG/L`
#> 
#> $`ENDRIN ALDEHYDE_TOTAL_NA_UG/L`
#> 
#> $`ENDRIN_TOTAL_NA_UG/L`
#> 
#> $`ESCHERICHIA COLI_NA_NA_#/100ML`
#> 
#> $`ESCHERICHIA COLI_NA_NA_MPN/100ML`
#> 
#> $FLOW_NA_NA_CFS
#> 
#> $`FLUORIDE_TOTAL_NA_UG/L`
#> 
#> $`GLYPHOSATE_DISSOLVED_NA_UG/L`
#> 
#> $`GLYPHOSATE_TOTAL_NA_UG/L`
#> 
#> $`HARDNESS, CA, MG_TOTAL_NA_UG/L`
#> 
#> $`HARDNESS, CARBONATE_NA_AS CACO3_UG/L`
#> 
#> $`HARDNESS, CARBONATE_NA_NA_UG/L`
#> 
#> $`HARDNESS, CARBONATE_TOTAL_AS CACO3_UG/L`
#> 
#> $`HARDNESS, CARBONATE_TOTAL_NA_UG/L`
#> 
#> $`ICE THICKNESS_NA_NA_IN`
#> 
#> $`IRON_DISSOLVED_NA_UG/L`
#> 
#> $`IRON_TOTAL_NA_UG/L`
#> 
#> $`KJELDAHL NITROGEN_TOTAL RECOVERABLE_AS N_UG/L`
#> 
#> $`KJELDAHL NITROGEN_TOTAL RECOVERABLE_NA_UG/L`
#> 
#> $`KJELDAHL NITROGEN_TOTAL_AS N_UG/L`
#> 
#> $`KJELDAHL NITROGEN_TOTAL_NA_UG/L`
#> 
#> $`LEAD_DISSOLVED_NA_UG/L`
#> 
#> $`LEAD_TOTAL_NA_UG/L`
#> 
#> $`MAGNESIUM_DISSOLVED_NA_UG/L`
#> 
#> $`MANGANESE_DISSOLVED_NA_UG/L`
#> 
#> $`MANGANESE_TOTAL_NA_UG/L`
#> 
#> $`MERCURY_DISSOLVED_NA_UG/L`
#> 
#> $`MERCURY_TOTAL_NA_UG/L`
#> 
#> $`MOLYBDENUM_DISSOLVED_NA_UG/L`
#> 
#> $`NICKEL_DISSOLVED_NA_UG/L`
#> 
#> $`NICKEL_TOTAL_NA_UG/L`
#> 
#> $`NITRATE + NITRITE_DISSOLVED_AS N_UG/L`
#> 
#> $`NITRATE + NITRITE_INORGANIC_AS NO3_UG/L`
#> 
#> $`NITRATE + NITRITE_TOTAL_AS N_UG/L`
#> 
#> $`NITRATE_TOTAL_AS NO3_UG/L`
#> 
#> $`NITRATE_TOTAL_AS N_UG/L`
#> 
#> $`NITRATE_TOTAL_NA_UG/L`
#> 
#> $`NITRATE_TOTAL_NONE_UG/L`
#> 
#> $`NITRATE_UNFILTERED_AS N_UG/L`
#> 
#> $`NITRITE_TOTAL_AS NO2_UG/L`
#> 
#> $`NITRITE_TOTAL_AS N_UG/L`
#> 
#> $`NITRITE_TOTAL_NA_UG/L`
#> 
#> $`NITRITE_UNFILTERED_AS N_UG/L`
#> 
#> $`NITROGEN_TOTAL_NONE_UG/L`
#> 
#> $`ORGANIC CARBON_DISSOLVED_NA_UG/L`
#> 
#> $`ORGANIC CARBON_TOTAL_NA_UG/L`
#> 
#> $`ORTHOPHOSPHATE_DISSOLVED_AS P_UG/L`
#> 
#> $`ORTHOPHOSPHATE_TOTAL RECOVERABLE_AS P_UG/L`
#> 
#> $`ORTHOPHOSPHATE_TOTAL_AS PO4_UG/L`
#> 
#> $`ORTHOPHOSPHATE_TOTAL_AS P_UG/L`
#> 
#> $`ORTHOPHOSPHATE_TOTAL_NA_UG/L`
#> 
#> $`PERIPHYTON_NA_NA_G/M2`
#> 
#> $`PHEOPHYTIN A_TOTAL_NA_UG/L`
#> 
#> $`PHOSPHORUS_TOTAL_AS PO4_UG/L`
#> 
#> $`PHOSPHORUS_TOTAL_AS P_UG/L`
#> 
#> $`PHOSPHORUS_TOTAL_NA_UG/L`
#> 
#> $PH_NA_NA_NONE
#> 
#> $`POTASSIUM_DISSOLVED_NA_UG/L`
#> 
#> $`PRESSURE_NA_NA_G/M2`
#> 
#> $PRESSURE_NA_NA_MMHG
#> 
#> $`RADIUM-226_NA_NA_PCI/L`
#> 
#> $`RADIUM-226_TOTAL_NA_PCI/L`
#> 
#> $`RADIUM-228_NA_NA_PCI/L`
#> 
#> $`RADIUM-228_TOTAL_NA_PCI/L`
#> 
#> $SALINITY_NA_NA_PSS
#> 
#> $`SELENIUM_DISSOLVED_NA_UG/L`
#> 
#> $`SELENIUM_TOTAL_NA_UG/L`
#> 
#> $`SILICON_TOTAL_NA_UG/L`
#> 
#> $`SILVER_DISSOLVED_NA_UG/L`
#> 
#> $`SODIUM_DISSOLVED_NA_UG/L`
#> 
#> $`SPECIFIC CONDUCTANCE_NA_NA_US/CM`
#> 
#> $`STREAM STAGE_NA_NA_FT`
#> 
#> $`SULFATE_TOTAL_AS SO4_UG/L`
#> 
#> $`SULFATE_TOTAL_NA_UG/L`
#> 
#> $`TEMPERATURE, AIR_NA_NA_DEG C`
#> 
#> $`TEMPERATURE, SAMPLE_NA_NA_DEG C`
#> 
#> $`TEMPERATURE, WATER_NA_NA_DEG C`
#> 
#> $`THALLIUM_DISSOLVED_NA_UG/L`
#> 
#> $`TOTAL DISSOLVED SOLIDS_NA_NA_UG/KG`
#> 
#> $`TOTAL DISSOLVED SOLIDS_NA_NA_UG/L`
#> 
#> $`TOTAL DISSOLVED SOLIDS_TOTAL_NA_UG/L`
#> 
#> $`TOTAL HARDNESS_TOTAL_NA_UG/L`
#> 
#> $`TOTAL KJELDAHL NITROGEN (ORGANIC N & NH3)_UNFILTERED_NA_UG/L`
#> 
#> $`TOTAL NITROGEN, MIXED FORMS_TOTAL_AS N_UG/L`
#> 
#> $`TOTAL NITROGEN, MIXED FORMS_TOTAL_NA_UG/L`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_NA_UG/L`
#> 
#> $`TOTAL SUSPENDED SOLIDS_NA_NA_UG/L`
#> 
#> $`TOXAPHENE_TOTAL_NA_UG/L`
#> 
#> $`TRANSPARENCY, SECCHI TUBE WITH DISK_NA_NA_CM`
#> 
#> $`TRITIUM_NA_NA_PCI/L`
#> 
#> $`TURBIDITY FIELD_NA_NA_FNU`
#> 
#> $`TURBIDITY FIELD_NA_NA_FTU`
#> 
#> $TURBIDITY_NA_NA_FNU
#> 
#> $TURBIDITY_NA_NA_FTU
#> 
#> $TURBIDITY_TOTAL_NA_FTU
#> 
#> $`URANIUM_DISSOLVED_NA_UG/L`
#> 
#> $`VANADIUM_DISSOLVED_NA_UG/L`
#> 
#> $`VOLATILE SUSPENDED SOLIDS_TOTAL_NA_UG/L`
#> 
#> $`ZINC_DISSOLVED_NA_UG/L`
#> 
#> $`ZINC_TOTAL_NA_UG/L`
#> 

# Create a single histogram using defaults. The input dataframe in this example
# is filtered so it includes only one TADA.ComparableDataIdentifier
df = dplyr::filter(Data_6Tribes_5y_Harmonized, TADA.ComparableDataIdentifier == "TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L")
TADA_Histogram(df, id_cols = "TADA.ComparableDataIdentifier")

# Create multiple histograms with additional grouping columns and view the first
# plot in list. In this example, we will group by both TADA.ComparableDataIdentifier
# and MonitoringLocationTypeName (e.g. stream, reservoir, canal, etc.)
# Load example dataset:
data(Data_Nutrients_Utah)
#> Warning: data set ‘Data_Nutrients_Utah’ not found
Histogram_output = TADA_Histogram(Data_Nutrients_UT, id_cols = c("TADA.ComparableDataIdentifier","MonitoringLocationTypeName"))
#> [1] "Plotting function removed 17 results where TADA.ResultMeasureValue = NA. These results cannot be plotted."
# This example generates 45 histograms
Histogram_output[[10]]

Histogram_output[[25]]

Histogram_output[[35]]