Create Scatterplot(s)

Create Scatterplot(s)

Usage

TADA_Scatterplot(.data, id_cols = c("TADA.ComparableDataIdentifier"))

Arguments

.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'TADA.ComparableDataIdentifier', 'TADA.ResultMeasureValue', and 'TADA.ResultMeasure.MeasureUnitCode' to run this function. The 'TADA.ComparableDataIdentifier' column is added to the data frame automatically when WQP data is retrieved using TADADataRetrieval. This TADA.ComparableDataIdentifier can be updated to harmonize synonyms by running the function TADA_HarmonizeSynonyms(). You can also include additional grouping columns in the id_cols input if desired. If more than one TADA.ComparableDataIdentifier exists in the dataset, the function creates a list of plots, where each list element name is a unique TADA.ComparableDataIdentifier.

id_cols

The column(s) in the dataset used to identify the unique groups to be plotted. Defaults to 'TADA.ComparableDataIdentifier'.

Value

A list of plotly scatterplot figures showing the distribution of sample values for each comparable data group (TADA.ComparableDataIdentifier).

Examples

# Load example dataset:
data(Data_6Tribes_5y_Harmonized)

# Create a scatterplot for each comparable data group (TADA.ComparableDataIdentifier)
# in the input dataframe:
TADA_Scatterplot(Data_6Tribes_5y_Harmonized, id_cols = "TADA.ComparableDataIdentifier")
#> $`ALKALINITY, TOTAL TOTAL AS CACO3 PCU`
#> 
#> $`ALPHA PARTICLE DISSOLVED PCI/L`
#> 
#> $`ALPHA PARTICLE PCI/L`
#> 
#> $`ALPHA PARTICLE TOTAL PCI/L`
#> 
#> $`ALUMINUM DISSOLVED UG/L`
#> 
#> $`AMMONIA FILTERED AS N UG/L`
#> 
#> $`AMMONIA UNFILTERED AS N UG/L`
#> 
#> $`AMMONIUM UNFILTERED AS N UG/L`
#> 
#> $`APPARENT COLOR TOTAL PCU`
#> 
#> $`ARSENIC DISSOLVED UG/L`
#> 
#> $`ARSENIC TOTAL UG/L`
#> 
#> $`BARIUM DISSOLVED UG/L`
#> 
#> $`BAROMETRIC PRESSURE G/M2`
#> 
#> $`BIOCHEMICAL OXYGEN DEMAND, STANDARD CONDITIONS TOTAL UG/L`
#> 
#> $`CADMIUM TOTAL UG/L`
#> 
#> $`CALCIUM DISSOLVED UG/L`
#> 
#> $`CHEMICAL OXYGEN DEMAND TOTAL UG/L`
#> 
#> $`CHLORIDE TOTAL UG/L`
#> 
#> $`CHLOROPHYLL A, CORRECTED FOR PHEOPHYTIN SUSPENDED UG/L`
#> 
#> $`CHLOROPHYLL A UG/L`
#> 
#> $`CHLOROPHYLL A UNFILTERED UG/L`
#> 
#> $`CHROMIUM TOTAL UG/L`
#> 
#> $`CLOUD COVER %`
#> 
#> $`CONDITION CLASS (DISSOLVED OXYGEN (DO)) %`
#> 
#> $`CONDUCTIVITY UG/L`
#> 
#> $`CONDUCTIVITY US/CM`
#> 
#> $`COPPER DISSOLVED UG/L`
#> 
#> $`COPPER TOTAL UG/L`
#> 
#> $`COUNT COUNT`
#> 
#> $`CYANIDE TOTAL UG/L`
#> 
#> $`DEPTH, BOTTOM IN`
#> 
#> $`DEPTH, SECCHI DISK DEPTH M`
#> 
#> $`DEPTH, SNOW COVER IN`
#> 
#> $`DEPTH IN`
#> 
#> $`DISCHARGE, RIVER/STREAM CFS`
#> 
#> $`DISSOLVED OXYGEN (DO) MG/L`
#> 
#> $`DISSOLVED OXYGEN SATURATION %`
#> 
#> $`DISSOLVED OXYGEN UPTAKE UG/L`
#> 
#> $`ESCHERICHIA COLI CFU/100ML`
#> 
#> $`FLOW CFS`
#> 
#> $`FLOW TOTAL CFS`
#> 
#> $`FLUORIDE TOTAL UG/L`
#> 
#> $FUNGI
#> 
#> $`HARDNESS, CA, MG TOTAL UG/L`
#> 
#> $`HARDNESS, CARBONATE AS CACO3 UG/L`
#> 
#> $`HARDNESS, CARBONATE UG/L`
#> 
#> $`HARDNESS, CARBONATE TOTAL AS CACO3 UG/L`
#> 
#> $`HARDNESS, CARBONATE TOTAL UG/L`
#> 
#> $`HEIGHT, GAGE IN`
#> 
#> $`ICE THICKNESS IN`
#> 
#> $`IRON DISSOLVED UG/L`
#> 
#> $`IRON TOTAL UG/L`
#> 
#> $`KJELDAHL NITROGEN TOTAL RECOVERABLE UG/L`
#> 
#> $`LEAD TOTAL UG/L`
#> 
#> $`LENGTH, TOTAL (FISH) IN`
#> 
#> $`LOONS, VISUAL OBSERVATION COUNT`
#> 
#> $`MAGNESIUM DISSOLVED UG/L`
#> 
#> $`MANGANESE DISSOLVED UG/L`
#> 
#> $`MANGANESE TOTAL UG/L`
#> 
#> $`MERCURY DISSOLVED UG/L`
#> 
#> $`MERCURY FILTERED, LAB UG/L`
#> 
#> $`MERCURY TOTAL UG/L`
#> 
#> $`MERCURY UNFILTERED UG/L`
#> 
#> $`METHYLMERCURY(1+) FILTERED, LAB UG/L`
#> 
#> $`METHYLMERCURY(1+) UNFILTERED UG/L`
#> 
#> $`NICKEL TOTAL UG/L`
#> 
#> $`NITRATE + NITRITE FILTERED AS N UG/L`
#> 
#> $`NITRATE + NITRITE INORGANIC AS NO3 UG/L`
#> 
#> $`NITRATE UNFILTERED AS N UG/L`
#> 
#> $`NITRATE UNFILTERED UG/L`
#> 
#> $`NITRITE UNFILTERED AS N UG/L`
#> 
#> $`NITRITE UNFILTERED UG/L`
#> 
#> $`NON-VIABLE SEED COUNT COUNT`
#> 
#> $`NON-VIABLE SEED WEIGHT G`
#> 
#> $`NUMBER OF SEEDS WITH ERGOTS COUNT`
#> 
#> $`NUMBER OF SEEDS WITH WORM HOLES COUNT`
#> 
#> $`NUMBER OF STALKS PER SAMPLE PLANT COUNT`
#> 
#> $`ORGANIC CARBON DISSOLVED UG/L`
#> 
#> $`ORGANIC CARBON TOTAL UG/L`
#> 
#> $`ORTHOPHOSPHATE FILTERED AS P UG/L`
#> 
#> $`ORTHOPHOSPHATE TOTAL RECOVERABLE AS P UG/L`
#> 
#> $`ORTHOPHOSPHATE UNFILTERED AS P UG/L`
#> 
#> $`PERIPHYTON G/M2`
#> 
#> $`PHEOPHYTIN A TOTAL UG/L`
#> 
#> $PH
#> 
#> $`PLANT HEIGHT (ABOVE WATER) IN`
#> 
#> $`PLANT HEIGHT (TOTAL) IN`
#> 
#> $`POTASSIUM DISSOLVED UG/L`
#> 
#> $`PRESSURE G/M2`
#> 
#> $`PRESSURE MMHG`
#> 
#> $`RADIUM-226 PCI/L`
#> 
#> $`RADIUM-226 TOTAL PCI/L`
#> 
#> $`RADIUM-228 PCI/L`
#> 
#> $`ROOT WEIGHT G`
#> 
#> $`SALINITY PSS`
#> 
#> $`SELENIUM TOTAL UG/L`
#> 
#> $`SHOOT WEIGHT G`
#> 
#> $`SILICON TOTAL UG/L`
#> 
#> $`SODIUM DISSOLVED UG/L`
#> 
#> $`SPECIFIC CONDUCTANCE, CALCULATED/MEASURED RATIO US/CM`
#> 
#> $`SPECIFIC CONDUCTANCE US/CM`
#> 
#> $`STREAM STAGE IN`
#> 
#> $`STREAM WIDTH MEASURE IN`
#> 
#> $`SULFATE TOTAL AS SO4 UG/L`
#> 
#> $`SULFATE TOTAL UG/L`
#> 
#> $`TEMPERATURE, AIR DEG C`
#> 
#> $`TEMPERATURE, SAMPLE DEG C`
#> 
#> $`TEMPERATURE DEG C`
#> 
#> $`TOTAL DISSOLVED SOLIDS UG/KG`
#> 
#> $`TOTAL DISSOLVED SOLIDS UG/L`
#> 
#> $`TOTAL DISSOLVED SOLIDS TOTAL UG/L`
#> 
#> $`TOTAL HARDNESS TOTAL UG/L`
#> 
#> $`TOTAL KJELDAHL NITROGEN (ORGANIC N & NH3) UNFILTERED AS N UG/L`
#> 
#> $`TOTAL KJELDAHL NITROGEN (ORGANIC N & NH3) UNFILTERED UG/L`
#> 
#> $`TOTAL NITROGEN, MIXED FORMS UNFILTERED AS N UG/L`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS FILTERED AS P UG/L`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L`
#> 
#> $`TOTAL SAMPLE WEIGHT G`
#> 
#> $`TOTAL SUSPENDED SOLIDS NON-FILTERABLE (PARTICLE) UG/L`
#> 
#> $`TRANSPARENCY, SECCHI TUBE WITH DISK IN`
#> 
#> $`TRITIUM PCI/L`
#> 
#> $`TRUE COLOR TOTAL PCU`
#> 
#> $`TURBIDITY FIELD NTU`
#> 
#> $`TURBIDITY NTU`
#> 
#> $`URANIUM DISSOLVED UG/L`
#> 
#> $`VIABLE SEED COUNT COUNT`
#> 
#> $`VIABLE SEED WEIGHT G`
#> 
#> $`VOLATILE SUSPENDED SOLIDS TOTAL UG/L`
#> 
#> $`WIND VELOCITY M/SEC`
#> 
#> $`ZINC DISSOLVED UG/L`
#> 
#> $`ZINC TOTAL UG/L`
#> 

# Create a single scatterplot using defaults. The input dataframe in this
# example is filtered so it includes only one TADA.ComparableDataIdentifier
df <- dplyr::filter(Data_6Tribes_5y_Harmonized, TADA.ComparableDataIdentifier == "TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L")
TADA_Scatterplot(df, id_cols = "TADA.ComparableDataIdentifier")

# Creates a scatterplot for each monitoring location
TADA_Scatterplot(df, id_cols = c("TADA.ComparableDataIdentifier", "MonitoringLocationName"))
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Alaska Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Ankeewinsee Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Aspen Ranch`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Balif Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bang (Long) Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bass Lake: Northwest Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bass Lake: Southeast Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Battle River Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Battle River North Branch`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Battle River South Branch`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bear Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Creek 001`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Lake (North Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Lake (South Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Thunder Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Traverse Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Blackduck River Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Blackduck River at Hwy 1 Bridge`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Blake Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Border Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Buffalo Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cahill Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cedar Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Chain: Middle Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Chain: South Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cimarron 002`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cimarron 003`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cimarron River`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Clearwater River CLWR-B`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Clearwater River at Boundary`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Colombo Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cormorant River at Hwy 23 Culverts`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Deadfish Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Dickens Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Dune Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L East Twin Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Elephant Ear Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Emerald Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Euchee 003`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fairbanks Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L First Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fond du Lac Creek Station 2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fox Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Francis Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fullers Lake: East Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fullers Lake: West Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Gibibwisher Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Gourd Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Grass Island Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Green Lake (Red Lake)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Green Lake (Redby)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Grouse Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Gwin Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Hardwood Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Harrison Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Hay Creek on Ponemah Rd`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Heart Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Infiltration Diversion/Upper Ponds`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Island Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Jaskari Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Joe Martin Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Johnson Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Jourdain Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Kesagiagan Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Kinney Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Big Traverse Bay Site 5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Four Mile Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Muskeg Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Near Long Point`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lac Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lake of the Woods at Big Traverse Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lake of the Woods at Central Big Traverse Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lake of the Woods at Muskeg Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Laxon Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Little Rock Creek at East Fireline Rd`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Little Thunder Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Long Lake (Burt)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lost Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : East`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : East Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : West`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake: West Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lussier Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Martin Branch (Marshall Rd.)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Miller (Mud) Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Morrison Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River Middle`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River at MUDR-M`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River at MUDR-U`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River at South Boundary Road`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River on East Fireline Trail`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Muerlin Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Navajo Pond`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L No-Name Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Otter Creek (Station 1)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pat Martin Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Perch Lake (North Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Perch Lake (South Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pickerel Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek Off Reservation`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at Barton's Camp Trail`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at PIKE-B`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at PIKE-I`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at South Boundary`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek on West Fireline Rd`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pine Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Poplar Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Poplar Creek by Upper Red Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Red Lake River at Lower Red Lake Outlet Dam`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Red Lake River at Outlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Red Lake River at Rock Dam`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Redhead Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Rice Portage Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Rio Chupadero`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Rio Tesuque`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Round Lake (Sylvia)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Rush Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L S007-884 SHOTLEY BK AT CSAH-23`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L SLRM 38.5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L SLRM 40.5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L SLRM 46.5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L SLRM 51`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L SLRM 53`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy River Upstream`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy River at Gonvick Truck Trail`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy River at SANR-U`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Second Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Seed Bank Well`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Shackle Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Shell Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Shemahgun Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Simian Creek Station 1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Simian Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Skull 002`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sofie Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L South Boundary`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Spruce (Spirit) Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Squaw Smith Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Stoney Brook Station 2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Stony Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (13')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (16')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (17')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (18')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (19')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (20')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Tamarac River at Steel Bridge`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Teal Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Tesuque Pueblo Bridge`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Third Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Third Lake (14')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Townline Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L USGS Guaging station Above Diversions/USFS Boundary`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Unamed Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: East`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: East Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: West`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: West-Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Veteran's Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Vigil Grant`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L West Twin Lake (North Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L West Twin Lake (South Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Wild Rice Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Williams Lake`
#> 

# Create multiple scatterplots with additional grouping columns and view the first
# plot in list. In this example, we will group by both TADA.ComparableDataIdentifier
# and MonitoringLocationTypeName (e.g. stream, reservoir, canal, etc.)
# Load example dataset:
data(Data_Nutrients_UT)
Scatterplot_output <- TADA_Scatterplot(Data_Nutrients_UT, id_cols = c("TADA.ComparableDataIdentifier", "MonitoringLocationTypeName"))
# This example generates 47 scatterplots
Scatterplot_output[[10]]

Scatterplot_output[[25]]

Scatterplot_output[[35]]