Generate Statistics Table

Generate Statistics Table

Usage

TADA_stats(.data)

Arguments

.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'TADA.ComparableDataIdentifier" and 'MonitoringLocationIdentifier' to run this function. The 'TADA.ComparableDataIdentifier' can be added to the data frame by running the function HarmonizeData().

Value

stats table

Examples

# Load example dataset:
data("TADAProfileClean18_TNonly")
# TADAProfileClean18_TNonly dataframe is clean, harmonized, and filtered
# down to one Comparable Data Identifier

# Create stats table:
TADA_stats(TADAProfileClean18_TNonly)
#> # A tibble: 1 × 17
#>   TADA.ComparableDataIdent…¹ n_sites n_records UpperFence LowerFence   Min  Mean
#>   <chr>                        <int>     <int>      <dbl>      <dbl> <dbl> <dbl>
#> 1 Total Nitrogen, Mixed For…     630      5005       2.59      -1.01 0.084  2.04
#> # ℹ abbreviated name: ¹​TADA.ComparableDataIdentifier
#> # ℹ 10 more variables: Max <dbl>, Percentile_5th <dbl>, Percentile_10th <dbl>,
#> #   Percentile_15th <dbl>, Percentile_25th <dbl>, Percentile_50th_Median <dbl>,
#> #   Percentile_75th <dbl>, Percentile_85th <dbl>, Percentile_95th <dbl>,
#> #   Percentile_98th <dbl>