TADA_ViewATTAINS

Visualizes the data returned from TADA_GetATTAINS if return_sf was set to TRUE.

Usage

TADA_ViewATTAINS(.data)

Arguments

.data

A list containing a data frame and ATTAINS shapefile objects created by TADA_GetATTAINS() with the return_sf argument set to TRUE.

Value

A leaflet map visualizing the TADA water quality observations and the linked ATTAINS assessment units. All maps are in WGS84.

Details

This function visualizes the raw ATTAINS features that are linked to the TADA Water Quality Portal observations. For the function to work properly, the input dataframe must be the list produced from TADA_GetATTAINS() with return_sf = TRUE. The map also displays the Water Quality Portal monitoring locations used to find the ATTAINS features. Check out the TADAModule2.Rmd for an example workflow.

See also

Examples

if (FALSE) {
tada_data <- TADA_DataRetrieval(
  startDate = "1990-01-01",
  endDate = "1995-12-31",
  characteristicName = "pH",
  statecode = "NV",
  applyautoclean = TRUE
)

attains_data <- TADA_GetATTAINS(tada_data, return_sf = TRUE)

TADA_ViewATTAINS(attains_data)
}