Create Boxplot

Create Boxplot

Usage

TADA_boxplot(filtered.data)

Arguments

filtered.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'TADA.ComparableDataIdentifier", 'TADA.ResultMeasureValue', and 'TADA.ResultMeasure.MeasureUnitCode' to run this function. These columns can be added to the data frame by running the function HarmonizeData(). The data frame must be filtered down to a single comparable data identifier to run this function.

Value

A plotly boxplot figure showing the median, 25th percentile, 75th percentile, upper fence, lower fence, minimum, maximum, and data outliers for the given comparable data identifier.

Examples

# Load example dataset:
data(Nutrients_Utah)

# Note for future: consider creating example dataset which has been cleaned

# Filter out quality control samples:
Utah_qc <- dplyr::filter(Nutrients_Utah, !(ActivityTypeCode %in% ActivityTypeCode[grepl("Quality",ActivityTypeCode)]))

# Run HarmonizeData function:
Utah_harmonized <- HarmonizeData(Utah_qc)

# Filter to 1 Comparable Data Identifier:
Utah_filtered <- dplyr::filter(Utah_harmonized, TADA.ComparableDataIdentifier == "Total Nitrogen, Mixed Forms_as N_mg/L")

# Create boxplot:
TADA_boxplot(Utah_filtered)