Generate TADA-compatible dataframe from WQP Data

Retrieve data from Water Quality Portal (WQP) and generate a TADA-compatible dataframe. Note that the inputs (e.g. project, organization, siteType) with the exceptions of endDate and startDate match the web service call format from the online WQP GUI. endDate and startDate match the format suggested in USGS's dataRetrieval package (endDate = "YYYY-MM-DD"), which is a more familiar date format for R users than the WQP GUI's endDateHi = "MM-DD-YYYY".

Usage

TADAdataRetrieval(
  startDate = "null",
  endDate = "null",
  countycode = "null",
  huc = "null",
  siteid = "null",
  siteType = "null",
  characteristicName = "null",
  characteristicType = "null",
  sampleMedia = "null",
  statecode = "null",
  organization = "null",
  project = "null",
  applyautoclean = TRUE
)

Arguments

startDate

Start Date string in the format YYYY-MM-DD, for example, "2020-01-01"

endDate

End Date string in the format YYYY-MM-DD, for example, "2020-01-01"

countycode

Code that identifies a county

huc

A numeric code denoting a hydrologic unit. Example: "04030202". Different size hucs can be entered.

siteid

Unique monitoring station identifier

siteType

Type of waterbody

characteristicName

Name of parameter

characteristicType

Groups of environmental measurements/parameters.

sampleMedia

Sampling substrate such as water, air, or sediment

statecode

Code that identifies a state

organization

A string of letters and/or numbers (some additional characters also possible) used to signify an organization with data in the Water Quality Portal

project

A string of letters and/or numbers (some additional characters also possible) used to signify a project with data in the Water Quality Portal

applyautoclean

Logical, defaults to TRUE. Applies TADA's autoclean function on the returned data profile.

Value

TADA-compatible dataframe

Details

This function will create and/or edit the following columns: TADA.DetectionLimitMeasureValue.Flag DetectionQuantitationLimitMeasure.MeasureValue DetectionLimitMeasureValue.Original ResultMeasureValue.Original TADA.ResultMeasureValue.Flag ResultMeasureValue

Keep in mind that all the query filters for the WQP work as an AND, but within the fields there are ORs. For example, within characteristicName – if you enter, c("pH", "Dissolved oxygen (DO)), – it’s an OR. Meaning that you will be back "pH" or "Dissolved oxygen (DO)" data. Similarly, if you enter c("Virginia", "Illinois"), it’s an OR. But the combo of fields are ANDs. Such as State/Virginia AND Characteristic/Dissolved oxygen (DO)". Meaning, you will get get back only Dissolved oxygen (DO) from Virginia. characteristicName and Characteristic Group also work as an AND, therefore the characteristicName must fall within the Characteristic Group when both are entered.

All data cleaning and transformations are done directly to the "ResultMeasureValue" and "DetectionLimitMeasureValue" columns, however the original "ResultMeasureValue" and "DetectionLimitMeasureValue" columns and values from the WQP are preserved in these new fields, "ResultMeasureValue.Original" and "DetectionLimitMeasureValue.Original". Additionally, "TADA.ResultMeasureValue.Flag" and "TADA.DetectionLimitMeasureValue.Flag" are created to track and changes made to the "ResultMeasureValue" and "DetectionLimitMeasureValue" columns; and to provide information about the result values that is needed to address censored data later on (i.e., nondetections)

Users can reference the WQX domain tables to find allowable values for queries, e.g., reference the WQX domain table to find countycode and statecode: https://cdx.epa.gov/wqx/download/DomainValues/County_CSV.zip Alternatively, you can use the WQP services to find areas where data is available in the US: https://www.waterqualitydata.us/Codes/countycode

See ?ConvertSpecialChars and ?autoclean documentation for more information.

Note: TADAdataRetrieval (by leveraging dataRetrieval), automatically converts the date times to UTC. It also automatically converts the data to dates, datetimes, numerics based on a standard algorithm. See: ?dataRetrieval::readWQPdata

Examples

if (FALSE) {
tada1 <- TADAdataRetrieval(statecode = "WI", countycode = "Dane", characteristicName = "Phosphorus")

tada2 <- TADAdataRetrieval(project = "Anchorage Bacteria 20-21")

tada3 <- TADAdataRetrieval(statecode = "UT", characteristicName = c("Ammonia", "Nitrate", "Nitrogen")) 

tada4 <- TADAdataRetrieval(statecode = "SC", countycode  = "Abbeville")

# example for CT
tada5 <- TADAdataRetrieval(statecode = "CT", startDate = "2020-10-01")

# note that countycode queries require a statecode (see example below)
tada6 <- TADAdataRetrieval(countycode = "US:02:020")

# example for NM
tada7 <- TADAdataRetrieval(statecode = "NM", 
                           characteristicName = c("Ammonia",
                                                  "Nitrate", 
                                                  "Nitrogen"), 
                           startDate = "2020-05-01")
                           
# example for AK
tada8 <- TADAdataRetrieval(project = "Anchorage Bacteria 20-21")

# another example for AK
tada9 <- TADAdataRetrieval(statecode = "AK", 
                           characteristicName = c("Fecal Coliform",
                                                  "Escherichia coli",
                                                  "Enterococcus", 
                                                  "Ammonia", 
                                                  "Nitrate", 
                                                  "Nitrogen"),
                           startDate = "2018-05-01")

# example for tribes
# Download data for many of the ATTAINS participating tribes
# Note this query may take about a half hour to run
# https://www.itecmembers.org/attains/
# ATTAINS participating tribes also have tribal pages in EPA's 
# How's My Waterway Application
# Example: https://mywaterway.epa.gov/tribe/SFNOES
#
# Sac & Fox Nation, Oklahoma "SFNOES_WQX"
# Citizen Potawatomi Nation, Oklahoma "CPNWATER"
# Delaware Nation, Oklahoma "DELAWARENATION"
# Hoopa Valley Tribe, California "HVTEPA_WQX"
# Otoe Missouria Tribe of Oklahoma "O_MTRIBE_WQX"
# Minnesota Chippewa Tribe, Minnesota (Fond du Lac Band) "FONDULAC_WQX"
# Pueblo of San Ildefonso, New Mexico "SANILDEFONSODECP"
# Pueblo of Santa Ana, New Mexico "PUEBLO_SANTAANA"
# Pueblo of Tesuque, New Mexico "PUEBLOOFTESUQUE"
# Red Lake Band of Chippewa Indians, Minnesota "REDLAKE_WQX"
# Seneca-Cayuga Nation "SCEQ"
# The Chickasaw Nation "CNENVSER"
# The Choctaw Nation of Oklahoma "CHOCNATWQX"
# Wyandotte Nation "WNENVDPT_WQX"
# Pueblo of Pojoaque "PUEBLO_POJOAQUE"

tada10 <- TADAdataRetrieval(organization = c("SFNOES_WQX",
                                             "CPNWATER",
                                             "DELAWARENATION",
                                             "HVTEPA_WQX",
                                             "O_MTRIBE_WQX",
                                             "FONDULAC_WQX",
                                             "SANILDEFONSODECP",
                                             "PUEBLO_SANTAANA",
                                             "PUEBLOOFTESUQUE",
                                             "REDLAKE_WQX",
                                             "SCEQ",
                                             "CNENVSER",
                                             "CHOCNATWQX",
                                             "WNENVDPT_WQX",
                                             "PUEBLO_POJOAQUE"))
}