Get text for tribal marker popup getPopup is used within TADA_addPolys and TADA_addPoints

Get text for tribal marker popup getPopup is used within TADA_addPolys and TADA_addPoints

Usage

getPopup(layer, layername)

Arguments

layer

A map feature layer

layername

Name of the layer

Value

Vector of strings to be used as the text for the popups when clicking on a tribal marker

Examples

# Get the Oklahoma Tribal Statistical Areas feature layer
layer <- TADA:::getFeatureLayer("https://geopub.epa.gov/arcgis/rest/services/EMEF/Tribal/MapServer/4/query")
# Get popup text for individual markers
TADA:::getPopup(layer, "Oklahoma Tribal Statistical Areas")
#>  [1] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Caddo Nation of Oklahoma<br><strong>EPA ID</strong>: 100000027<br>"                                                    
#>  [2] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Wichita and Affiliated Tribes (Wichita, Keechi, Waco, & Tawakonie), Oklahoma<br><strong>EPA ID</strong>: 100000325<br>"
#>  [3] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Delaware Nation, Oklahoma<br><strong>EPA ID</strong>: 100000072<br>"                                                   
#>  [4] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Cherokee Nation<br><strong>EPA ID</strong>: 100000038<br>"                                                             
#>  [5] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Cheyenne and Arapaho Tribes, Oklahoma<br><strong>EPA ID</strong>: 100000039<br>"                                       
#>  [6] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: The Chickasaw Nation<br><strong>EPA ID</strong>: 100000041<br>"                                                        
#>  [7] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: The Choctaw Nation of Oklahoma<br><strong>EPA ID</strong>: 100000045<br>"                                              
#>  [8] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Citizen Potawatomi Nation, Oklahoma<br><strong>EPA ID</strong>: 100000046<br>"                                         
#>  [9] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Absentee-Shawnee Tribe of Indians of Oklahoma<br><strong>EPA ID</strong>: 100000001<br>"                               
#> [10] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: The Muscogee (Creek) Nation<br><strong>EPA ID</strong>: 100000169<br>"                                                 
#> [11] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Eastern Shawnee Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000077<br>"                                           
#> [12] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Iowa Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000112<br>"                                                      
#> [13] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Kaw Nation, Oklahoma<br><strong>EPA ID</strong>: 100000122<br>"                                                        
#> [14] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Kickapoo Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000127<br>"                                                  
#> [15] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Kiowa Indian Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000129<br>"                                              
#> [16] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Comanche Nation, Oklahoma<br><strong>EPA ID</strong>: 100000052<br>"                                                   
#> [17] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Apache Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000007<br>"                                                    
#> [18] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Fort Sill Apache Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000092<br>"                                          
#> [19] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Miami Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000158<br>"                                                     
#> [20] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Modoc Nation<br><strong>EPA ID</strong>: 100000164<br>"                                                                
#> [21] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Otoe-Missouria Tribe of Indians, Oklahoma<br><strong>EPA ID</strong>: 100000187<br>"                                   
#> [22] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Ottawa Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000186<br>"                                                    
#> [23] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Pawnee Nation of Oklahoma<br><strong>EPA ID</strong>: 100000197<br>"                                                   
#> [24] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Peoria Tribe of Indians of Oklahoma<br><strong>EPA ID</strong>: 100000200<br>"                                         
#> [25] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Ponca Tribe of Indians of Oklahoma<br><strong>EPA ID</strong>: 100000206<br>"                                          
#> [26] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Quapaw Nation<br><strong>EPA ID</strong>: 100000230<br>"                                                               
#> [27] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Sac & Fox Nation, Oklahoma<br><strong>EPA ID</strong>: 100000248<br>"                                                  
#> [28] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: The Seminole Nation of Oklahoma<br><strong>EPA ID</strong>: 100000265<br>"                                             
#> [29] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Seneca-Cayuga Nation<br><strong>EPA ID</strong>: 100000268<br>"                                                        
#> [30] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Tonkawa Tribe of Indians of Oklahoma<br><strong>EPA ID</strong>: 100000304<br>"                                        
#> [31] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Wyandotte Nation<br><strong>EPA ID</strong>: 100000330<br>"                                                            
#> [32] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: The Muscogee (Creek) Nation<br><strong>EPA ID</strong>: 100000169<br>"                                                 
#> [33] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: The Seminole Nation of Oklahoma<br><strong>EPA ID</strong>: 100000265<br>"                                             
#> [34] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Kaw Nation, Oklahoma<br><strong>EPA ID</strong>: 100000122<br>"                                                        
#> [35] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Ponca Tribe of Indians of Oklahoma<br><strong>EPA ID</strong>: 100000206<br>"                                          
#> [36] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Kiowa Indian Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000129<br>"                                              
#> [37] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Comanche Nation, Oklahoma<br><strong>EPA ID</strong>: 100000052<br>"                                                   
#> [38] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Apache Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000007<br>"                                                    
#> [39] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Fort Sill Apache Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000092<br>"                                          
#> [40] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Caddo Nation of Oklahoma<br><strong>EPA ID</strong>: 100000027<br>"                                                    
#> [41] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Wichita and Affiliated Tribes (Wichita, Keechi, Waco, & Tawakonie), Oklahoma<br><strong>EPA ID</strong>: 100000325<br>"
#> [42] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Delaware Nation, Oklahoma<br><strong>EPA ID</strong>: 100000072<br>"                                                   
#> [43] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Miami Tribe of Oklahoma<br><strong>EPA ID</strong>: 100000158<br>"                                                     
#> [44] "<strong>Oklahoma Tribal Statistical Areas</strong><p><strong>Tribe Name</strong>: Peoria Tribe of Indians of Oklahoma<br><strong>EPA ID</strong>: 100000200<br>"