Generate list of field names subset by parameter

Function subsets the input dataframe by the input parameter and creates a list of the fields in the subset dataframe as well as the number of unique values in each field. The list is intended to inform users as to which specific fields to explore further and filter (subset by a parameter).

Usage

FilterParFields(.data, parameter)

Arguments

.data

TADA dataframe

parameter

Characteristic name (parameter name) from the dataframe.

Value

A table of fields and the count of unique values in each field, subset by a parameter.

Details

Recommended fields to review and filter by for each parameter: "ActivityCommentText" "ActivityTypeCode" "TADA.ActivityMediaName" "ActivityMediaSubdivisionName" "MeasureQualifierCode" "MonitoringLocationTypeName" "HydrologicCondition" "HydrologicEvent" "ResultStatusIdentifier" "MethodQualifierTypeName" "ResultCommentText" "ResultLaboratoryCommentText" "TADA.ResultMeasure.MeasureUnitCode" "TADA.ResultSampleFractionText" "ResultTemperatureBasisText" "ResultValueTypeName" "ResultWeightBasisText" "SampleCollectionEquipmentName" "LaboratoryName" "MethodDescriptionText" "ResultParticleSizeBasisText" "SampleCollectionMethod.MethodIdentifier" "SampleCollectionMethod.MethodIdentifierContext" "SampleCollectionMethod.MethodName" "DataQuality.BiasValue" "MethodSpeciationName" "ResultAnalyticalMethod.MethodName" "ResultAnalyticalMethod.MethodIdentifier" "ResultAnalyticalMethod.MethodIdentifierContext" "AssemblageSampledName" "DetectionQuantitationLimitTypeName"

Examples

# Load example dataset:
data(Nutrients_Utah)

# Create list of fields for parameter "AMMONIA" with number of unique values in each field:
AmmoniaFields <- FilterParFields(Nutrients_Utah, parameter = "AMMONIA")
#>                                            Fields Count
#> 1                                ActivityTypeCode     5
#> 2                            MeasureQualifierCode     3
#> 3                                  LaboratoryName     3
#> 4              DetectionQuantitationLimitTypeName     3
#> 5                      MonitoringLocationTypeName     3
#> 6         SampleCollectionMethod.MethodIdentifier     2
#> 7  SampleCollectionMethod.MethodIdentifierContext     2
#> 8               SampleCollectionMethod.MethodName     2
#> 9                            MethodSpeciationName     2
#> 10                  TADA.ResultSampleFractionText     2
#> 11                   ActivityMediaSubdivisionName     1
#> 12                            ActivityCommentText     1
#> 13                            HydrologicCondition     1
#> 14                                HydrologicEvent     1
#> 15                  SampleCollectionEquipmentName     1
#> 16                         ResultStatusIdentifier     1
#> 17                            ResultValueTypeName     1
#> 18                          ResultWeightBasisText     1
#> 19                     ResultTemperatureBasisText     1
#> 20                    ResultParticleSizeBasisText     1
#> 21                          DataQuality.BiasValue     1
#> 22                              ResultCommentText     1
#> 23        ResultAnalyticalMethod.MethodIdentifier     1
#> 24 ResultAnalyticalMethod.MethodIdentifierContext     1
#> 25              ResultAnalyticalMethod.MethodName     1
#> 26                    ResultLaboratoryCommentText     1
#> 27                         TADA.ActivityMediaName     1
#> 28             TADA.ResultMeasure.MeasureUnitCode     1