Large WQP data pulls using dataRetrieval

This function does multiple synchronous data calls to the WQP (waterqualitydata.us). It uses the WQP summary service to limit the amount downloaded to only relevant data, and pulls back data from 100 stations at a time and then joins the data back together and produces a single TADA compatible dataframe as the output. For large data sets, that can save a lot of time and ultimately reduce the complexity of subsequent data processing. Using this function, you will be able to download all data available from all sites in the contiguous United States that is available for the time period, characteristicName, and siteType requested. Computer memory may limit the size of datasets that your R console will be able to hold in one session. Function requires a characteristicName, siteType, statecode, huc, or start/ end date input. The recommendation is to be as specific as you can with your large data call. The function allows the user to run autoclean on the dataset, but this is not the default as checking large dataframes for exact duplicate rows can be time consuming and is better performed on its own once the query is completed.

Usage

TADABigdataRetrieval(
  startDate = "null",
  endDate = "null",
  huc = "null",
  siteid = "null",
  siteType = "null",
  characteristicName = "null",
  characteristicType = "null",
  sampleMedia = "Water",
  statecode = "null",
  maxrecs = 250000,
  applyautoclean = FALSE
)

Arguments

startDate

Start Date string in the format YYYY-MM-DD, for example, "2020-01-01"

endDate

End Date string in the format YYYY-MM-DD, for example, "2020-01-01"

huc

A numeric code denoting a hydrologic unit. Example: "04030202". Different size hucs can be entered.

siteid

Unique monitoring station identifier

siteType

Type of waterbody

characteristicName

Name of parameter

characteristicType

Groups of environmental measurements/parameters.

sampleMedia

Sampling substrate such as water, air, or sediment

statecode

Code that identifies a state

maxrecs

The maximum number of results queried within one call to dataRetrieval.

applyautoclean

Logical, defaults to TRUE. Applies TADA's autoclean function on the returned data profile.

Value

TADA-compatible dataframe

Details

Some code for this function was adapted from this USGS Blog (Author: Aliesha Krall) Large Sample Pull

See ?autoclean documentation for more information on this optional input.

Note: TADABigdataRetrieval (by leveraging dataRetrieval), automatically converts the date times to UTC. It also automatically converts the data to dates, datetimes, numerics based on a standard algorithm. See: ?dataRetrieval::readWQPdata

Examples

if (FALSE) {
tada1 <- TADABigdataRetrieval(startDate = "2019-01-01", endDate = "2021-12-31", characteristicName = "Temperature, water", statecode = c("AK","AL"))
tada2 <- TADABigdataRetrieval(startDate = "2016-10-01",endDate = "2022-09-30", statecode = "UT")
tada3 = TADABigdataRetrieval(huc = "04030202", characteristicName = "Escherichia coli")
tada4 = TADABigdataRetrieval(huc = c("04030202","04030201"), characteristicName = "Temperature, water")
}