Create Boxplot(s)

Create Boxplot(s)

Usage

TADA_Boxplot(.data, id_cols = c("TADA.ComparableDataIdentifier"))

Arguments

.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'TADA.ComparableDataIdentifier', 'TADA.ResultMeasureValue', and 'TADA.ResultMeasure.MeasureUnitCode' to run this function. 'TADA.ComparableDataIdentifier' can be added to the data frame by running the function HarmonizeData(). The user can include additional grouping columns in the id_col input. If more than one group exists in the dataset (i.e. two or more unique comparable data identifiers), the function creates a list of plots, where each list element name is a unique group identifier.

id_cols

The column(S) in the dataset used to identify the unique groups to be plotted. Defaults to 'TADA.ComparableDataIdentifier'.

Value

A list of plotly boxplot figures showing the median, 25th percentile, 75th percentile, upper fence, lower fence, minimum, maximum, and data outliers for each unique data group.

Examples

# Create a single boxplot using defaults. The input dataframe in this example
# includes only one unique TADA.ComparableDataIdentifier: 
# Load example dataset:
data(Data_6Tribes_5y_Harmonized)
# Filter data down to a single TADA.ComparableDataIdentifier
df = dplyr::filter(Data_6Tribes_5y_Harmonized, TADA.ComparableDataIdentifier == "TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L")
# Generate single boxplot
TADA_Boxplot(df, id_cols = "TADA.ComparableDataIdentifier")

# Create multiple boxplots with additional grouping columns and view the first
# plot in list. In this example, we will group data in the input dataframe
# by both the TADA.ComparableDataIdentifier and the OrganizationIdentifier
Boxplots_TPbyOrg = TADA_Boxplot(df, id_cols = c("TADA.ComparableDataIdentifier","OrganizationIdentifier"))
# This example generates 4 box plots.
Boxplots_TPbyOrg[[1]]
#> $visdat
#> $visdat$`1bb17b29ef6d`
#> function () 
#> plotlyVisDat
#> <bytecode: 0x563764580b60>
#> <environment: 0x56377ad47f50>
#> 
#> 
#> $cur_data
#> [1] "1bb17b29ef6d"
#> 
#> $attrs
#> $attrs$`1bb17b29ef6d`
#> $attrs$`1bb17b29ef6d`$y
#> $attrs$`1bb17b29ef6d`$y[[1]]
#>  [1]  24.13140  41.74080  37.82760  25.10970  38.47980  90.98190  71.08980
#>  [8]  29.67510  41.41470  55.76310 154.89750 183.59430  53.80650  27.39240
#> [15]  25.10970 511.97700 258.27120 176.42010  60.32850  34.89270  18.91380
#> [22]  95.87340  40.43640  28.37070  29.34900  30.97950 298.70760 260.88000
#> [29] 117.72210  32.61000  24.45750  80.87280 450.01800 348.92700 368.49300
#> [36]  77.28570  21.61487  88.71775  20.00182  57.42458  36.45493  28.71229
#> [43]  27.74446  27.42185  25.76190  16.30500  25.10970  66.19830  44.67570
#> [50]  38.80590  40.11030  14.02230  20.54430  44.67570  18.26160  21.52260
#> [57] 335.88300 134.35320  43.04520 198.59490  38.47980  21.52260  34.56660
#> [64] 146.74500 100.76490  67.17660  27.42185  25.10970  13.04400  26.74020
#> [71]  28.37070  41.08860
#> 
#> 
#> $attrs$`1bb17b29ef6d`$q1
#>    25% 
#> 27.229 
#> 
#> $attrs$`1bb17b29ef6d`$median
#> [1] 39.4581
#> 
#> $attrs$`1bb17b29ef6d`$q3
#>    75% 
#> 89.284 
#> 
#> $attrs$`1bb17b29ef6d`$lowerfence
#> [1] 13.044
#> 
#> $attrs$`1bb17b29ef6d`$hoverinfo
#> [1] "y"
#> 
#> $attrs$`1bb17b29ef6d`$upperfence
#> [1] 176.4201
#> 
#> $attrs$`1bb17b29ef6d`$boxpoints
#> [1] "outliers"
#> 
#> $attrs$`1bb17b29ef6d`$marker
#> $attrs$`1bb17b29ef6d`$marker$color
#> [1] "#00bde3"
#> 
#> 
#> $attrs$`1bb17b29ef6d`$stroke
#> [1] "#005ea2"
#> 
#> $attrs$`1bb17b29ef6d`$alpha_stroke
#> [1] 1
#> 
#> $attrs$`1bb17b29ef6d`$sizes
#> [1]  10 100
#> 
#> $attrs$`1bb17b29ef6d`$spans
#> [1]  1 20
#> 
#> $attrs$`1bb17b29ef6d`$type
#> [1] "box"
#> 
#> 
#> 
#> $layout
#> $layout$width
#> NULL
#> 
#> $layout$height
#> NULL
#> 
#> $layout$margin
#> $layout$margin$b
#> [1] 40
#> 
#> $layout$margin$l
#> [1] 60
#> 
#> $layout$margin$t
#> [1] 25
#> 
#> $layout$margin$r
#> [1] 10
#> 
#> 
#> 
#> $source
#> [1] "A"
#> 
#> $config
#> $config$modeBarButtonsToAdd
#> [1] "hoverclosest" "hovercompare"
#> 
#> $config$showSendToCloud
#> [1] FALSE
#> 
#> $config$displayModeBar
#> [1] FALSE
#> 
#> 
#> $layoutAttrs
#> $layoutAttrs$`1bb17b29ef6d`
#> $layoutAttrs$`1bb17b29ef6d`$xaxis
#> $layoutAttrs$`1bb17b29ef6d`$xaxis$showticklabels
#> [1] FALSE
#> 
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$title
#> [1] "UG/L"
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$titlefont
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$titlefont$size
#> [1] 16
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$titlefont$family
#> [1] "Arial"
#> 
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$tickfont
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$tickfont$size
#> [1] 16
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$tickfont$family
#> [1] "Arial"
#> 
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$hoverformat
#> [1] ",.4r"
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$linecolor
#> [1] "black"
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$rangemode
#> [1] "tozero"
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$showgrid
#> [1] FALSE
#> 
#> $layoutAttrs$`1bb17b29ef6d`$yaxis$tickcolor
#> [1] "black"
#> 
#> 
#> $layoutAttrs$`1bb17b29ef6d`$hoverlabel
#> $layoutAttrs$`1bb17b29ef6d`$hoverlabel$bgcolor
#> [1] "white"
#> 
#> 
#> $layoutAttrs$`1bb17b29ef6d`$title
#> [1] "Boxplot of \nTOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L_PUEBLOOFTESUQUE"
#> 
#> $layoutAttrs$`1bb17b29ef6d`$plot_bgcolor
#> [1] "#e5ecf6"
#> 
#> $layoutAttrs$`1bb17b29ef6d`$margin
#> $layoutAttrs$`1bb17b29ef6d`$margin$l
#> [1] 50
#> 
#> $layoutAttrs$`1bb17b29ef6d`$margin$r
#> [1] 20
#> 
#> $layoutAttrs$`1bb17b29ef6d`$margin$b
#> [1] 20
#> 
#> $layoutAttrs$`1bb17b29ef6d`$margin$t
#> [1] 55
#> 
#> $layoutAttrs$`1bb17b29ef6d`$margin$pad
#> [1] 0
#> 
#> 
#> 
#> 
#> attr(,"TOJSON_FUNC")
#> function (x, ...) 
#> {
#>     jsonlite::toJSON(x, digits = 50, auto_unbox = TRUE, force = TRUE, 
#>         null = "null", na = "null", time_format = "%Y-%m-%d %H:%M:%OS6", 
#>         ...)
#> }
#> <bytecode: 0x56376449e4a0>
#> <environment: namespace:plotly>
Boxplots_TPbyOrg[[2]]
#> NULL
Boxplots_TPbyOrg[[3]]
#> NULL
 
# Create multiple boxplots with additional grouping columns and view the first
# plot in list. In this example, we will group data in the input dataframe
# by both the TADA.ComparableDataIdentifier and the MonitoringLocationTypeName 
# (e.g. stream, reservoir, canal, etc.)
# Load example dataset:
data(Data_Nutrients_Utah)
#> Warning: data set ‘Data_Nutrients_Utah’ not found
Boxplot_output = TADA_Boxplot(Data_Nutrients_UT, id_cols = c("TADA.ComparableDataIdentifier","MonitoringLocationTypeName"))
#> [1] "Plotting function removed 17 results where TADA.ResultMeasureValue = NA. These results cannot be plotted."
# This example generates 45 box plots.
Boxplot_output[[2]]

Boxplot_output[[25]]

Boxplot_output[[40]]