Create Histogram(s)

Create Histogram(s)

Usage

TADA_Histogram(.data, id_cols = c("TADA.ComparableDataIdentifier"))

Arguments

.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'TADA.ComparableDataIdentifier', 'TADA.ResultMeasureValue', and 'TADA.ResultMeasure.MeasureUnitCode' to run this function. 'TADA.ComparableDataIdentifier' can be added to the data frame by running the function HarmonizeData(). The user can include additional grouping columns in the id_col input. If more than one group exists in the dataset (i.e. two or more unique comparable data identifiers), the function creates a list of plots, where each list element name is a unique group identifier.

id_cols

The column(S) in the dataset used to identify the unique groups to be plotted. Defaults to 'TADA.ComparableDataIdentifier'.

Value

A list of plotly histogram figures showing the distribution of sample values for each data group.

Examples

# Create histogram using defaults, and view first plot in list:
# Load example dataset:
data(Data_TP_6Tribes_5y)

Histogram_output = TADA_Histogram(Data_TP_6Tribes_5y, id_cols = "TADA.ComparableDataIdentifier")
Histogram_output[[1]]
#> $visdat
#> $visdat$`1bb46ff76a0d`
#> function () 
#> plotlyVisDat
#> <bytecode: 0x555ccafa3018>
#> <environment: 0x555cdb2477c0>
#> 
#> 
#> $cur_data
#> [1] "1bb46ff76a0d"
#> 
#> $attrs
#> $attrs$`1bb46ff76a0d`
#> $attrs$`1bb46ff76a0d`$alpha_stroke
#> [1] 1
#> 
#> $attrs$`1bb46ff76a0d`$sizes
#> [1]  10 100
#> 
#> $attrs$`1bb46ff76a0d`$spans
#> [1]  1 20
#> 
#> 
#> $attrs$`1bb46ff76a0d`
#> $alpha_stroke
#> [1] 1
#> 
#> $sizes
#> [1]  10 100
#> 
#> $spans
#> [1]  1 20
#> 
#> $x
#>    [1]  10.0  17.0  10.0  16.0  13.0  14.0  21.0  15.0  12.0  15.0  11.0  15.0
#>   [13]  21.0  18.0  21.0  16.0  20.0  27.0  27.0  27.0  27.0  15.0  15.0  18.0
#>   [25]  23.0  24.0  32.0  10.0  15.0  13.0  10.0  10.0  46.0  53.0  27.0  23.0
#>   [37]  47.0  46.0  32.0  48.0  20.0  20.0  17.0  44.0  27.0  43.0  47.0  24.0
#>   [49]  22.0  41.0  20.0  45.0  47.0  13.0  13.0  15.0  16.0  20.0  14.0  17.0
#>   [61]  11.0  12.0  16.0  13.0  15.0  12.0  11.0   9.0   6.0  10.0  11.0  13.0
#>   [73]  21.0  54.0  26.0  40.0  34.0  20.0  51.0  39.0  34.0  35.0  21.0  14.0
#>   [85]  14.0  12.0  12.0  58.0  17.0  10.0  10.0  10.0  11.0  14.0  16.0  10.0
#>   [97]   5.0   9.0   9.0   9.0   9.0   9.0  11.0   9.0  12.0  14.0  12.0   8.0
#>  [109]  12.0   7.0   8.0  10.0  12.0   7.0  32.0  44.0  18.0  62.0  50.0  39.0
#>  [121]  42.0  32.0  32.0  18.0  21.0  15.0  14.0  22.0  30.0  14.0  11.0  12.0
#>  [133]  10.0   9.0   8.0   9.0  12.0  10.0  10.0   6.0   9.0  14.0  12.0  13.0
#>  [145]  12.0   8.0  11.0  42.0  11.0  11.0   9.0   7.0  39.0  31.0  42.0  17.0
#>  [157]  37.0  34.0  12.0  25.0  14.0  20.0  12.0  10.0  21.0  36.0  16.0  40.0
#>  [169]  39.0  37.0  40.0  30.0  64.0  26.0  35.0  41.0  39.0  29.0  39.0  29.0
#>  [181]  38.0  71.0  33.0  31.0  17.0  44.0  34.0  39.0  32.0  63.0  28.0  18.0
#>  [193]  45.0  32.0  41.0  21.0  12.0  17.0  45.0  13.0  11.0  12.0  10.0  12.0
#>  [205]  20.0  25.0  20.0  16.0  20.0  12.0  16.0  21.0  15.0  15.0  81.0  17.0
#>  [217]  17.0  19.0  24.0  16.0  28.0   7.0  16.0  12.0  15.0   9.0  17.0  41.0
#>  [229]  13.0  18.0  38.0  46.0  19.0  43.0  31.0  21.0  36.0  20.0  17.0  18.0
#>  [241]  23.0  22.0  21.0  14.0  10.0  21.0  16.0  16.0  13.0  24.0  23.0  37.0
#>  [253]  20.0  36.0  34.0  35.0  34.0  35.0  35.0  41.0  40.0  41.0  35.0  25.0
#>  [265]  20.0  16.0  12.0  15.0  11.0  19.0  15.0   8.0  18.0  18.0  13.0  26.0
#>  [277]  13.0  21.0  14.0  17.0  23.0  21.0  19.0  19.0   9.0   8.0   9.0  20.0
#>  [289]  17.0  20.0  12.0  21.0  16.0  31.0  28.0  60.0  63.0  29.0  34.0  38.0
#>  [301]  32.0  29.0  31.0  36.0  36.0  34.0  31.0  36.0  54.0  51.0  40.0  42.0
#>  [313]  47.0  36.0  50.0  46.0  52.0  48.0  42.0  48.0  44.0  10.0  12.0  15.0
#>  [325]   8.0  19.0  13.0  20.0  12.0  14.0  16.0  17.0  16.0  13.0  16.0  12.0
#>  [337]   8.0   9.0  13.0  15.0   9.0  28.0  11.0   6.0  12.0  21.0  13.0  12.0
#>  [349]  11.0  12.0   7.0  10.0  17.0  40.0  39.0  42.0  39.0  24.0  29.0  25.0
#>  [361]  24.0  26.0  27.0  24.0  33.0  31.0  33.0  27.0  33.0  32.0  20.0  46.0
#>  [373]  24.0   9.0  12.0  11.0  14.0  20.0  11.0   6.0  11.0  10.0  26.0  13.0
#>  [385]   8.0  14.0   7.0  11.0   9.0   6.0  11.0   8.0   8.0  15.0  18.0  16.0
#>  [397]  13.0  12.0  13.0  14.0  10.0  10.0  11.0   8.0  10.0  37.0  44.0  33.0
#>  [409]  45.0  46.0  50.0  51.0  10.0  11.0  14.0  16.0  13.0  12.0   9.0  14.0
#>  [421]  22.0  10.0  49.0  46.0  47.0  53.0  48.0  48.0  36.0  50.0  46.0  51.0
#>  [433]  12.0  18.0  11.0   6.0  16.0  15.0  19.0  11.0  12.0  20.0  14.0  24.0
#>  [445]  11.0  18.0  11.0   6.0  12.0  14.0  11.0  46.0  13.0  10.0  21.0  12.0
#>  [457]   8.0  14.0  12.0   8.0  22.0   8.0   9.0  10.0  48.0  63.0  57.0  60.0
#>  [469] 124.0  71.0  82.0   4.0  11.0  14.0  13.0 210.0  16.0  22.0   9.0  22.0
#>  [481]   8.0  11.0  20.0   8.0  14.0  12.0  10.0  11.0  26.0   7.0  20.0  11.0
#>  [493]  10.0  16.0  23.0  24.0  23.0  28.0  25.0  35.0  38.0  16.0  16.0  16.0
#>  [505]  15.0  20.0  18.0  20.0  18.0  20.0  13.0  15.0  19.0  23.0  20.0  20.0
#>  [517]  21.0  20.0  18.0  10.0  15.0  16.0  14.0  27.0   9.0  18.0  13.0  14.0
#>  [529]  18.0  18.0  14.0  11.0  16.0  25.0  34.0  45.0  29.0  56.0  26.0  61.0
#>  [541]  45.0  35.0  48.0  24.0  21.0  32.0  28.0   9.0  10.0  13.0  14.0  14.0
#>  [553]  12.0  13.0   9.0  12.0   7.0  10.0  13.0  10.0  20.0   8.0  15.0  10.0
#>  [565]  10.0  12.0   9.0   8.0  12.0   8.0  12.0  39.0  39.0  41.0  44.0  38.0
#>  [577]  47.0  51.0  48.0  46.0  43.0  34.0  37.0  47.0  46.0  36.0  36.0  36.0
#>  [589]  38.0  35.0  50.0  51.0  50.0  50.0  58.0  10.0  12.0  12.0  14.0  14.0
#>  [601]  16.0  12.0  10.0  13.0   5.0  11.0  10.0  11.0  22.0   6.0  11.0   8.0
#>  [613]   9.0  10.0   8.0   7.0  10.0   8.0   9.0  53.0  54.0  56.0  32.0  35.0
#>  [625]  35.0  35.0  33.0  37.0  31.0  38.0  42.0  33.0  67.0  42.0  40.0  34.0
#>  [637]  42.0  15.0  11.0  13.0  12.0   6.0   8.0  17.0  21.0  18.0  16.0  12.0
#>  [649]  17.0  10.0  13.0  11.0  10.0  66.0  10.0  13.0  13.0   9.0   9.0   9.0
#>  [661]  14.0  34.0  35.0  30.0  58.0  54.0  16.0  16.0  15.0  13.0  15.0  17.0
#>  [673]  15.0  15.0  16.0  18.0   9.0  14.0  15.0  23.0  21.0  11.0  12.0  10.0
#>  [685]  11.0   9.0   8.0  10.0  14.0  19.0   8.0  28.0   7.0  23.0  11.0  18.0
#>  [697]  14.0  15.0  11.0   8.0  15.0  14.0  14.0  13.0  13.0  14.0  13.0  16.0
#>  [709]  14.0  15.0  12.0   5.0   8.0  14.0  11.0  28.0   8.0  13.0   8.0  11.0
#>  [721]   6.0   7.0   8.0  24.0  12.0   7.0   8.0  10.0   9.0   9.0   6.0   5.0
#>  [733]  48.0  14.0   8.0   8.0  29.0  15.0  19.0  18.0  17.0  18.0  24.0  21.0
#>  [745]  24.0  22.0  35.0  38.0  43.0  37.0  18.0  17.0  16.0  18.0  18.0  35.0
#>  [757]  35.0  31.0  41.0  30.0  29.0  34.0  26.0  39.0  38.0  35.0  41.0  39.0
#>  [769]  33.0  29.0  25.0  34.0  28.0  55.0  68.0  56.0  58.0  56.0  12.0  13.0
#>  [781]  20.0  22.0  13.0  10.0  10.0  11.0   6.0   9.0   8.0  10.0  14.0  10.0
#>  [793]   8.0   5.0  12.0   9.0   5.0  10.0  10.0  12.0  11.0  24.0  24.0  15.0
#>  [805] 109.0  50.0  47.0  27.0  28.0  32.0  27.0  29.0  48.0  59.0  47.0  39.0
#>  [817]  42.0   9.0   9.0  10.0  13.0  14.0  13.0  12.0   7.0   9.0   8.0  12.0
#>  [829]   9.0   9.0  25.0   7.0   9.0  15.0  12.0   7.0   8.0  13.0  12.0  12.0
#>  [841]  31.0  28.0  33.0  30.0  30.0  39.0  35.0  37.0  35.0  37.0   9.0  13.0
#>  [853]  14.0  18.0  17.0  23.0  14.0  11.0  17.0   9.0  13.0  11.0  13.0  36.0
#>  [865]   8.0  16.0   8.0  11.0   6.0  10.0  12.0  20.0  10.0  41.0  40.0  44.0
#>  [877]  43.0  40.0  48.0  49.0  47.0  48.0  48.0  42.0  43.0  39.0  37.0  42.0
#>  [889]  45.0  54.0  51.0  48.0  46.0  54.0  11.0  13.0  12.0  19.0  16.0  18.0
#>  [901]  12.0  15.0  14.0   7.0  10.0  10.0  12.0  45.0  10.0  14.0   8.0  14.0
#>  [913]  12.0   7.0   8.0  11.0  13.0  13.0  34.0  38.0  39.0  44.0  38.0  42.0
#>  [925]  34.0  39.0  38.0  39.0  31.0  37.0 287.0  30.0  31.0  34.0  35.0  35.0
#>  [937]  32.0  31.0   5.0   5.0   5.0  11.0  29.0  21.0  13.0  16.0  33.0  11.0
#>  [949]  11.0  12.0  17.0  13.0  13.0   7.9  12.0  15.0  18.0  17.0  12.0   6.9
#>  [961]  23.0  24.0  36.0  34.0  10.0  11.0  12.0  17.0  13.0  13.0   7.9  25.0
#>  [973]  15.0  18.0  15.0  43.0  32.0  43.0  33.0  37.0  40.0  44.0  41.0  30.0
#>  [985]  43.0  36.0  48.0  33.0  12.0  12.0   9.9  10.0  16.0  14.0  12.0  14.0
#>  [997]  20.0  12.0  10.0  15.0  15.0   9.2  12.0  10.0  14.0  13.0  12.0  15.0
#> [1009]  11.0  10.0   8.4  12.0  12.0  16.0   8.1  10.0  12.0   9.4  10.0  38.0
#> [1021]  45.0  50.0  30.0  78.0  53.0  49.0  36.0  20.0  66.0  41.0  42.0  51.0
#> [1033]  44.0  28.0  32.0  53.0  61.0  18.0  25.0  28.0  34.0  29.0  64.0  49.0
#> [1045]  30.0  36.0  28.0  37.0  50.0  30.0  39.0  50.0  75.0  20.0  22.0  18.0
#> [1057]  45.0  23.0  15.0  17.0  22.0  15.0  17.0  25.0  25.0  18.0   7.4   8.6
#> [1069]   8.6  11.0   9.9  22.0  10.0  12.0   9.9  16.0   8.4   8.3   8.2   9.6
#> [1081]   7.4  11.0  15.0  15.0  16.0  15.0  13.0  20.0  16.0  12.0  14.0  21.0
#> [1093]  14.0  11.0  10.0   8.0  18.0  16.0  20.0  17.0  41.0  36.0  87.0  26.0
#> [1105]  30.0  16.0  27.0  22.0  19.0  26.0  26.0  47.0  20.0  55.0  43.0  22.0
#> [1117]  35.0  41.0  32.0  22.0  31.0  67.0  17.0  15.0  15.0  15.0  27.0  42.0
#> [1129]  17.0  18.0  18.0  15.0  15.0  17.0  20.0  15.0  15.0  14.0  11.0  14.0
#> [1141]  15.0  11.0   9.3  15.0  15.0  14.0  17.0  30.0  29.0  21.0  14.0  24.0
#> [1153]  18.0  17.0  23.0  18.0  21.0  23.0  23.0  17.0  45.0  14.0  31.0  15.0
#> [1165]  24.0  14.0  33.0  39.0 120.0  51.0  20.0  27.0  22.0  20.0  32.0  19.0
#> [1177]  20.0  23.0  35.0  35.0  21.0  36.0  35.0  24.0  23.0  35.0  79.0  56.0
#> [1189]  50.0  51.0  75.0  82.0  97.0 190.0  50.0  34.0  20.0  22.0  21.0  32.0
#> [1201]  30.0  20.0  28.0  36.0  27.0  46.0  22.0  24.0  19.0  50.0  46.0  42.0
#> [1213]  24.0  39.0  25.0  28.0  27.0  24.0  25.0  45.0  37.0  39.0  42.0  33.0
#> [1225]  39.0  64.0  36.0  28.0  23.0  59.0  38.0  27.0  54.0  21.0  26.0  28.0
#> [1237]  22.0  19.0  27.0  38.0  21.0  21.0  37.0  25.0  19.0  22.0  25.0  32.0
#> [1249]   8.9  17.0  20.0  22.0  18.0  20.0  64.0  27.0  29.0  28.0  27.0  28.0
#> [1261]  41.0  26.0  13.0  10.0   9.5   7.4  26.0  15.0  14.0   6.9  14.0  11.0
#> [1273]  10.0  11.0  12.0  12.0  11.0  12.0  15.0  18.0  18.0  20.0  15.0  19.0
#> [1285]  17.0  18.0  16.0  19.0  20.0  18.0  22.0  18.0  16.0  15.0  11.0  16.0
#> [1297]  15.0  15.0  19.0  17.0  20.0  22.0  23.0  21.0  13.0  16.0  14.0  19.0
#> [1309]  15.0  12.0  15.0   9.4  12.0  44.0  16.0  12.0   9.9  21.0  14.0  21.0
#> [1321]  13.0  56.0 102.0  69.0 119.0 109.0 108.0 118.0  82.0 145.0  97.0  82.0
#> [1333]  68.0  74.0  62.0  48.0  90.0  46.0  59.0  58.0  94.0 124.0 129.0  87.0
#> [1345]  62.0  94.0  91.0  62.0 154.0 161.0  22.0  60.0  34.0  56.0  34.0 402.0
#> [1357] 160.0  76.0  35.0  24.0  38.0  38.0  43.0  38.0  64.0  76.0 102.0 117.0
#> [1369] 109.0  56.0 160.0  65.0  76.0  74.0  90.0 111.0 109.0  80.0  86.0  55.0
#> [1381]  53.0  73.0  60.0  49.0  52.0  46.0  39.0  36.0  29.0  13.0  13.0  13.0
#> [1393]  16.0  14.0  16.0  14.0  15.0  15.0  21.0  15.0  13.0  14.0   9.0   6.0
#> [1405]   7.0  16.0  15.0   7.0  23.0  12.0  16.0   9.0  10.0  24.0  13.0   7.0
#> [1417]   6.0  17.0  13.0   9.0   7.0   7.0  87.0  71.0 256.0 158.0 141.0  69.0
#> [1429]  16.0  52.0 381.0 263.0  53.0  55.0  71.0  44.0  59.0  47.0  55.0  38.0
#> [1441]  30.0  31.0  32.0  35.0  37.0  32.0  27.0  36.0  39.0  56.0  39.0  40.0
#> [1453]  37.0  42.0  30.0  37.0  41.0  37.0  33.0  38.0  37.0  30.0  37.0  28.0
#> [1465]  34.0  31.0  34.0  31.0  36.0  34.0  39.0  44.0  40.0  45.0  38.0  49.0
#> [1477]  38.0  47.0  52.0  49.0  16.0  18.0  12.0  15.0  13.0  15.0  15.0   8.0
#> [1489]   9.0  16.0  10.0  25.0   6.0  10.0  12.0  14.0   9.0   8.0   8.0  13.0
#> [1501]  10.0  43.0  40.0  80.0  66.0  61.0  73.0  33.0  56.0  51.0  62.0 101.0
#> [1513]  45.0  47.0  69.0  57.0   9.0  12.0  15.0   8.0   9.0   9.0   7.0   9.0
#> [1525]   8.0   8.0  25.0   6.0  11.0   8.0  14.0  10.0   6.0   8.0  12.0  10.0
#> [1537]   8.0  34.0  36.0  52.0  42.0  10.0   9.0  14.0   3.0   9.0   9.0   7.0
#> [1549]   8.0  10.0   9.0  20.0   9.0  10.0  10.0  18.0  10.0   6.0   8.0   8.0
#> [1561]   9.0  41.0  41.0  43.0  35.0  35.0  50.0  44.0  38.0  44.0  39.0  49.0
#> [1573]  47.0  44.0  49.0  49.0  43.0  74.0  69.0  41.0  40.0  40.0  39.0  35.0
#> [1585]  34.0  38.0  42.0  42.0  47.0  36.0  31.0  29.0  32.0  26.0  29.0  64.0
#> [1597]  46.0  40.0  39.0  39.0  52.0  30.0  25.0  30.0  34.0  52.0  51.0  55.0
#> [1609]  51.0  61.0  59.0  43.0  50.0  42.0  51.0  56.0   8.0  10.0  11.0  14.0
#> [1621]  16.0  11.0   8.0  11.0   8.0   9.0  10.0  12.0  32.0   8.0  12.0   8.0
#> [1633]  10.0  11.0   8.0   6.0  10.0   8.0   8.0  59.0  54.0  51.0  54.0  57.0
#> [1645]  55.0  49.0  55.0  53.0  50.0  45.0  14.0  15.0  13.0  14.0  12.0  11.0
#> [1657]  13.0  11.0   9.0  10.0   6.0   7.0  18.0   9.0  23.0   6.0   5.0   4.0
#> [1669]   6.0   5.0   8.0   7.0  14.0   8.0   7.0   6.0   7.0  18.0   5.0   6.0
#> [1681]  12.0   5.0  13.0  36.0  41.0  14.0  17.0  14.0  13.0  14.0  14.0  12.0
#> [1693]  14.0  14.0  12.0  93.0  84.0 122.0  64.0 117.0  54.0  54.0  55.0 122.0
#> [1705]  73.0  39.0  66.0  81.0  43.0  25.0  91.0  92.0  48.0  68.0  37.0   6.0
#> [1717]  40.0  51.0  29.0  28.0  38.0  24.0  63.0  50.0 102.0  38.0  36.0  38.0
#> [1729]  59.0  46.0 151.0  59.0 175.0  15.0  62.0  86.0  32.0  78.0   1.5 112.0
#> [1741]  48.0  62.0  45.0  53.0  73.0  74.0  83.0   1.5  38.0  49.0  25.0  75.0
#> [1753]  50.0  28.0  47.0  30.0  21.0   1.5  60.0  25.0  70.0  89.0 102.0  32.0
#> [1765] 109.0  29.0  40.0  21.0  42.0  57.0  81.0  41.0  43.0  16.0 155.0  33.0
#> [1777]  46.0  55.0  52.0  74.0  37.0  46.0  41.0  21.0  37.0  47.0  16.0  37.0
#> [1789]  27.0  26.0  32.0  31.0  21.0  37.0  80.0  61.0  76.0  97.0  39.0  40.0
#> [1801]  20.0  33.0  53.0  37.0  40.0  17.0 184.0  24.0 110.0  94.0  54.0  42.0
#> [1813]  18.0  27.0  14.0  49.0  33.0  60.0  21.0  32.0  35.0  40.0  24.0  43.0
#> [1825]  81.0  63.0  36.0  15.0   1.5  73.0  90.0  46.0  32.0  38.0  35.0  38.0
#> [1837]  23.0  33.0  39.0  70.0  21.0  41.0  33.0  63.0  37.0  46.0  38.0  33.0
#> [1849]  41.0  27.0  20.0  27.0  55.0  52.0  24.0  57.0  20.0  61.0  45.0  73.0
#> [1861]  37.0  73.0  34.0  26.0  35.0  48.0  47.0  10.0  34.0  46.0  29.0  44.0
#> [1873]  47.0  65.0  70.0  38.0  31.0  26.0  48.0  15.0  54.0  29.0  52.0  80.0
#> [1885]  24.0  43.0  38.0  37.0  25.0  18.0  46.0  71.0  73.0  54.0  36.0  37.0
#> [1897]  42.0  43.0  30.0  52.0  36.0  42.0  62.0  65.0  91.0  57.0  38.0   1.5
#> [1909]  30.0  67.0 107.0  42.0  65.0  38.0  30.0  46.0  41.0  29.0  50.0  46.0
#> [1921]  62.0  21.0  72.0  13.0  76.0  28.0  59.0  21.0  27.0  25.0  49.0  43.0
#> [1933]  36.0  63.0  83.0  50.0  58.0  30.0  14.0  61.0  30.0  31.0  10.0  45.0
#> [1945] 123.0  24.0  39.0  42.0  54.0  26.0  50.0  39.0 207.0  37.0  38.0  57.0
#> [1957]  67.0  35.0  35.0  60.0  37.0  46.0  22.0  63.0  29.0  23.0  29.0  61.0
#> [1969] 147.0  52.0  36.0  25.0  11.0  34.0  44.0  30.0  54.0  38.0  68.0  80.0
#> [1981]  25.0  31.0  25.0  46.0  48.0  25.0   8.0  32.0  39.0  60.0  27.0  31.0
#> [1993]  43.0  36.0  72.0  46.0  38.0  47.0  50.0  58.0  17.0  57.0  62.0  62.0
#> [2005]  32.0 159.0  31.0  48.0  50.0 112.0  31.0 128.0  50.0  64.0  58.0  28.0
#> [2017]  40.0  11.0  39.0  35.0  67.0  35.0  41.0  41.0  41.0  27.0  14.0  26.0
#> [2029]  49.0  14.0  42.0  87.0  24.0  36.0  39.0  24.0  34.0  46.0  41.0  54.0
#> [2041]  52.0  25.0  59.0  33.0  38.0  21.0  41.0  61.0  46.0  21.0  45.0  45.0
#> [2053]  42.0  37.0  58.0  43.0  51.0  51.0  40.0  41.0  54.0  25.0  20.0  24.0
#> [2065]  33.0  49.0  32.0  10.0  54.0  33.0  42.0  10.0  48.0  10.0  33.0  39.0
#> [2077]  66.0  68.0  38.0  26.0 104.0  66.0  28.0  70.0  32.0  48.0  47.0 125.0
#> [2089]  21.0  32.0  41.0  26.0  23.0  48.0  44.0  56.0  61.0  45.0  36.0  51.0
#> [2101]  26.0  39.0  35.0  30.0  45.0  62.0  11.0  43.0  38.0  24.0  42.0  26.0
#> [2113] 134.0 218.0  41.0  25.0  56.0  49.0  49.0  59.0   1.5  39.0  38.0  50.0
#> [2125]  55.0  55.0  62.0  28.0  34.0  29.0  26.0  46.0  53.0  42.0  32.0  43.0
#> [2137]  28.0  39.0  33.0  25.0  58.0  55.0  53.0  35.0  52.0  61.0  22.0  34.0
#> [2149]  30.0  36.0  41.0  50.0  23.0  67.0  42.0  59.0  19.0  28.0  23.0  41.0
#> [2161]  26.0  40.0  47.0  37.0  26.0  70.0  40.0  38.0  32.0  42.0  56.0  58.0
#> [2173]  76.0  36.0  40.0  35.0   9.0  24.0  35.0  99.0  38.0 133.0  31.0  45.0
#> [2185]  53.0  60.0  22.0  23.0  37.0  48.0 108.0  57.0  40.0  50.0  84.0  14.0
#> [2197]  46.0  32.0  73.0  64.0   1.5  28.0  40.0  64.0  33.0  56.0  25.0  46.0
#> [2209]  41.0  29.0  50.0  77.0  69.0   8.0  43.0  71.0  28.0  39.0  23.0  76.0
#> [2221]  36.0  37.0  37.0  39.0  17.0  22.0  76.0  34.0  37.0  25.0  64.0  40.0
#> [2233]  37.0  45.0  83.0  26.0  26.0  44.0  42.0  54.0   6.0  53.0  47.0  35.0
#> [2245]   1.5  48.0  30.0  32.0  56.0  44.0 112.0  35.0  56.0  83.0  23.0  38.0
#> [2257]  32.0  23.0  40.0  21.0  65.0  24.0  99.0  65.0  46.0  71.0  33.0  43.0
#> [2269]  54.0  49.0  64.0  60.0  31.0  24.0  30.0  60.0  41.0  31.0  23.0  58.0
#> [2281]  25.0  24.0  30.0  22.0  93.0  18.0  36.0  26.0  22.0  15.0  38.0  37.0
#> [2293]  32.0  45.0  16.0   8.0  36.0  33.0  71.0  76.0  40.0  32.0  64.0  61.0
#> [2305]  44.0  43.0  22.0  75.0  98.0  55.0  25.0  65.0  35.0  42.0  50.0  35.0
#> [2317]  52.0   1.5 104.0  19.0  33.0  17.0   1.5   8.0  57.0 123.0   1.5  63.0
#> [2329]  30.0  45.0   1.5  55.0  33.0  65.0  55.0  21.0  18.0  85.0  39.0  57.0
#> [2341]  56.0  36.0  32.0  37.0  46.0  39.0  66.0  72.0  34.0  41.0  41.0  32.0
#> [2353] 194.0  62.0  24.0  21.0  33.0  44.0  36.0  99.0  98.0  51.0 123.0  34.0
#> [2365]  70.0  34.0  26.0  31.0  27.0  42.0  31.0  33.0  41.0  48.0  22.0  71.0
#> [2377]  66.0  47.0  60.0  48.0 157.0  33.0  51.0  96.0  15.0  65.0  63.0  32.0
#> [2389]  26.0  21.0  48.0  47.0  31.0   9.0  65.0  38.0  42.0  57.0  92.0  40.0
#> [2401]  21.0  25.0  54.0  41.0  15.0  58.0  62.0  35.0 101.0  12.0  43.0  28.0
#> [2413]  83.0  41.0  42.0  62.0  38.0  58.0  33.0  51.0  46.0  49.0  24.0  67.0
#> [2425]  39.0  42.0  36.0  80.0   1.5  24.0  23.0  22.0  33.0  17.0  36.0  13.0
#> [2437]  45.0  60.0  62.0 238.0  57.0  15.0  49.0  23.0  61.0  81.0  31.0  30.0
#> [2449]  36.0  38.0  35.0  43.0  42.0  35.0  48.0  34.0  95.0  33.0  41.0  23.0
#> [2461]  34.0  86.0  32.0  61.0  22.0  86.0  55.0  95.0  60.0  46.0  64.0  37.0
#> [2473]  22.0  46.0  22.0  35.0  39.0  42.0 148.0  40.0  89.0  52.0  50.0  54.0
#> [2485] 118.0 131.0  30.0  38.0  28.0  43.0  14.0  42.0 116.0  49.0  53.0  67.0
#> [2497]  59.0  64.0
#> 
#> $type
#> [1] "histogram"
#> 
#> $xbins
#> $xbins$start
#> [1] 1.5
#> 
#> 
#> $marker
#> $marker$color
#> [1] "#00bde3"
#> 
#> 
#> $stroke
#> [1] "#005ea2"
#> 
#> $bingroup
#> [1] 1
#> 
#> $name
#> [1] "<b>All Data<b>"
#> 
#> $inherit
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "plotly_histogram" "list"            
#> 
#> $attrs$`1bb46ff76a0d`
#> $alpha_stroke
#> [1] 1
#> 
#> $sizes
#> [1]  10 100
#> 
#> $spans
#> [1]  1 20
#> 
#> $x
#>    [1] 10.0 17.0 10.0 16.0 13.0 14.0 21.0 15.0 12.0 15.0 11.0 15.0 21.0 18.0
#>   [15] 21.0 16.0 20.0 27.0 27.0 27.0 27.0 15.0 15.0 18.0 23.0 24.0 32.0 10.0
#>   [29] 15.0 13.0 10.0 10.0 46.0 53.0 27.0 23.0 47.0 46.0 32.0 48.0 20.0 20.0
#>   [43] 17.0 44.0 27.0 43.0 47.0 24.0 22.0 41.0 20.0 45.0 47.0 13.0 13.0 15.0
#>   [57] 16.0 20.0 14.0 17.0 11.0 12.0 16.0 13.0 15.0 12.0 11.0  9.0  6.0 10.0
#>   [71] 11.0 13.0 21.0 54.0 26.0 40.0 34.0 20.0 51.0 39.0 34.0 35.0 21.0 14.0
#>   [85] 14.0 12.0 12.0 58.0 17.0 10.0 10.0 10.0 11.0 14.0 16.0 10.0  5.0  9.0
#>   [99]  9.0  9.0  9.0  9.0 11.0  9.0 12.0 14.0 12.0  8.0 12.0  7.0  8.0 10.0
#>  [113] 12.0  7.0 32.0 44.0 18.0 62.0 50.0 39.0 42.0 32.0 32.0 18.0 21.0 15.0
#>  [127] 14.0 22.0 30.0 14.0 11.0 12.0 10.0  9.0  8.0  9.0 12.0 10.0 10.0  6.0
#>  [141]  9.0 14.0 12.0 13.0 12.0  8.0 11.0 42.0 11.0 11.0  9.0  7.0 39.0 31.0
#>  [155] 42.0 17.0 37.0 34.0 12.0 25.0 14.0 20.0 12.0 10.0 21.0 36.0 16.0 40.0
#>  [169] 39.0 37.0 40.0 30.0 64.0 26.0 35.0 41.0 39.0 29.0 39.0 29.0 38.0 71.0
#>  [183] 33.0 31.0 17.0 44.0 34.0 39.0 32.0 63.0 28.0 18.0 45.0 32.0 41.0 21.0
#>  [197] 12.0 17.0 45.0 13.0 11.0 12.0 10.0 12.0 20.0 25.0 20.0 16.0 20.0 12.0
#>  [211] 16.0 21.0 15.0 15.0 81.0 17.0 17.0 19.0 24.0 16.0 28.0  7.0 16.0 12.0
#>  [225] 15.0  9.0 17.0 41.0 13.0 18.0 38.0 46.0 19.0 43.0 31.0 21.0 36.0 20.0
#>  [239] 17.0 18.0 23.0 22.0 21.0 14.0 10.0 21.0 16.0 16.0 13.0 24.0 23.0 37.0
#>  [253] 20.0 36.0 34.0 35.0 34.0 35.0 35.0 41.0 40.0 41.0 35.0 25.0 20.0 16.0
#>  [267] 12.0 15.0 11.0 19.0 15.0  8.0 18.0 18.0 13.0 26.0 13.0 21.0 14.0 17.0
#>  [281] 23.0 21.0 19.0 19.0  9.0  8.0  9.0 20.0 17.0 20.0 12.0 21.0 16.0 31.0
#>  [295] 28.0 60.0 63.0 29.0 34.0 38.0 32.0 29.0 31.0 36.0 36.0 34.0 31.0 36.0
#>  [309] 54.0 51.0 40.0 42.0 47.0 36.0 50.0 46.0 52.0 48.0 42.0 48.0 44.0 10.0
#>  [323] 12.0 15.0  8.0 19.0 13.0 20.0 12.0 14.0 16.0 17.0 16.0 13.0 16.0 12.0
#>  [337]  8.0  9.0 13.0 15.0  9.0 28.0 11.0  6.0 12.0 21.0 13.0 12.0 11.0 12.0
#>  [351]  7.0 10.0 17.0 40.0 39.0 42.0 39.0 24.0 29.0 25.0 24.0 26.0 27.0 24.0
#>  [365] 33.0 31.0 33.0 27.0 33.0 32.0 20.0 46.0 24.0  9.0 12.0 11.0 14.0 20.0
#>  [379] 11.0  6.0 11.0 10.0 26.0 13.0  8.0 14.0  7.0 11.0  9.0  6.0 11.0  8.0
#>  [393]  8.0 15.0 18.0 16.0 13.0 12.0 13.0 14.0 10.0 10.0 11.0  8.0 10.0 37.0
#>  [407] 44.0 33.0 45.0 46.0 50.0 51.0 10.0 11.0 14.0 16.0 13.0 12.0  9.0 14.0
#>  [421] 22.0 10.0 49.0 46.0 47.0 53.0 48.0 48.0 36.0 50.0 46.0 51.0 12.0 18.0
#>  [435] 11.0  6.0 16.0 15.0 19.0 11.0 12.0 20.0 14.0 24.0 11.0 18.0 11.0  6.0
#>  [449] 12.0 14.0 11.0 46.0 13.0 10.0 21.0 12.0  8.0 14.0 12.0  8.0 22.0  8.0
#>  [463]  9.0 10.0 48.0 63.0 57.0 60.0 71.0 82.0  4.0 11.0 14.0 13.0 16.0 22.0
#>  [477]  9.0 22.0  8.0 11.0 20.0  8.0 14.0 12.0 10.0 11.0 26.0  7.0 20.0 11.0
#>  [491] 10.0 16.0 23.0 24.0 23.0 28.0 25.0 35.0 38.0 16.0 16.0 16.0 15.0 20.0
#>  [505] 18.0 20.0 18.0 20.0 13.0 15.0 19.0 23.0 20.0 20.0 21.0 20.0 18.0 10.0
#>  [519] 15.0 16.0 14.0 27.0  9.0 18.0 13.0 14.0 18.0 18.0 14.0 11.0 16.0 25.0
#>  [533] 34.0 45.0 29.0 56.0 26.0 61.0 45.0 35.0 48.0 24.0 21.0 32.0 28.0  9.0
#>  [547] 10.0 13.0 14.0 14.0 12.0 13.0  9.0 12.0  7.0 10.0 13.0 10.0 20.0  8.0
#>  [561] 15.0 10.0 10.0 12.0  9.0  8.0 12.0  8.0 12.0 39.0 39.0 41.0 44.0 38.0
#>  [575] 47.0 51.0 48.0 46.0 43.0 34.0 37.0 47.0 46.0 36.0 36.0 36.0 38.0 35.0
#>  [589] 50.0 51.0 50.0 50.0 58.0 10.0 12.0 12.0 14.0 14.0 16.0 12.0 10.0 13.0
#>  [603]  5.0 11.0 10.0 11.0 22.0  6.0 11.0  8.0  9.0 10.0  8.0  7.0 10.0  8.0
#>  [617]  9.0 53.0 54.0 56.0 32.0 35.0 35.0 35.0 33.0 37.0 31.0 38.0 42.0 33.0
#>  [631] 67.0 42.0 40.0 34.0 42.0 15.0 11.0 13.0 12.0  6.0  8.0 17.0 21.0 18.0
#>  [645] 16.0 12.0 17.0 10.0 13.0 11.0 10.0 66.0 10.0 13.0 13.0  9.0  9.0  9.0
#>  [659] 14.0 34.0 35.0 30.0 58.0 54.0 16.0 16.0 15.0 13.0 15.0 17.0 15.0 15.0
#>  [673] 16.0 18.0  9.0 14.0 15.0 23.0 21.0 11.0 12.0 10.0 11.0  9.0  8.0 10.0
#>  [687] 14.0 19.0  8.0 28.0  7.0 23.0 11.0 18.0 14.0 15.0 11.0  8.0 15.0 14.0
#>  [701] 14.0 13.0 13.0 14.0 13.0 16.0 14.0 15.0 12.0  5.0  8.0 14.0 11.0 28.0
#>  [715]  8.0 13.0  8.0 11.0  6.0  7.0  8.0 24.0 12.0  7.0  8.0 10.0  9.0  9.0
#>  [729]  6.0  5.0 48.0 14.0  8.0  8.0 29.0 15.0 19.0 18.0 17.0 18.0 24.0 21.0
#>  [743] 24.0 22.0 35.0 38.0 43.0 37.0 18.0 17.0 16.0 18.0 18.0 35.0 35.0 31.0
#>  [757] 41.0 30.0 29.0 34.0 26.0 39.0 38.0 35.0 41.0 39.0 33.0 29.0 25.0 34.0
#>  [771] 28.0 55.0 68.0 56.0 58.0 56.0 12.0 13.0 20.0 22.0 13.0 10.0 10.0 11.0
#>  [785]  6.0  9.0  8.0 10.0 14.0 10.0  8.0  5.0 12.0  9.0  5.0 10.0 10.0 12.0
#>  [799] 11.0 24.0 24.0 15.0 50.0 47.0 27.0 28.0 32.0 27.0 29.0 48.0 59.0 47.0
#>  [813] 39.0 42.0  9.0  9.0 10.0 13.0 14.0 13.0 12.0  7.0  9.0  8.0 12.0  9.0
#>  [827]  9.0 25.0  7.0  9.0 15.0 12.0  7.0  8.0 13.0 12.0 12.0 31.0 28.0 33.0
#>  [841] 30.0 30.0 39.0 35.0 37.0 35.0 37.0  9.0 13.0 14.0 18.0 17.0 23.0 14.0
#>  [855] 11.0 17.0  9.0 13.0 11.0 13.0 36.0  8.0 16.0  8.0 11.0  6.0 10.0 12.0
#>  [869] 20.0 10.0 41.0 40.0 44.0 43.0 40.0 48.0 49.0 47.0 48.0 48.0 42.0 43.0
#>  [883] 39.0 37.0 42.0 45.0 54.0 51.0 48.0 46.0 54.0 11.0 13.0 12.0 19.0 16.0
#>  [897] 18.0 12.0 15.0 14.0  7.0 10.0 10.0 12.0 45.0 10.0 14.0  8.0 14.0 12.0
#>  [911]  7.0  8.0 11.0 13.0 13.0 34.0 38.0 39.0 44.0 38.0 42.0 34.0 39.0 38.0
#>  [925] 39.0 31.0 37.0 30.0 31.0 34.0 35.0 35.0 32.0 31.0  5.0  5.0  5.0 11.0
#>  [939] 29.0 21.0 13.0 16.0 33.0 11.0 11.0 12.0 17.0 13.0 13.0  7.9 12.0 15.0
#>  [953] 18.0 17.0 12.0  6.9 23.0 24.0 36.0 34.0 10.0 11.0 12.0 17.0 13.0 13.0
#>  [967]  7.9 25.0 15.0 18.0 15.0 43.0 32.0 43.0 33.0 37.0 40.0 44.0 41.0 30.0
#>  [981] 43.0 36.0 48.0 33.0 12.0 12.0  9.9 10.0 16.0 14.0 12.0 14.0 20.0 12.0
#>  [995] 10.0 15.0 15.0  9.2 12.0 10.0 14.0 13.0 12.0 15.0 11.0 10.0  8.4 12.0
#> [1009] 12.0 16.0  8.1 10.0 12.0  9.4 10.0 38.0 45.0 50.0 30.0 78.0 53.0 49.0
#> [1023] 36.0 20.0 66.0 41.0 42.0 51.0 44.0 28.0 32.0 53.0 61.0 18.0 25.0 28.0
#> [1037] 34.0 29.0 64.0 49.0 30.0 36.0 28.0 37.0 50.0 30.0 39.0 50.0 75.0 20.0
#> [1051] 22.0 18.0 45.0 23.0 15.0 17.0 22.0 15.0 17.0 25.0 25.0 18.0  7.4  8.6
#> [1065]  8.6 11.0  9.9 22.0 10.0 12.0  9.9 16.0  8.4  8.3  8.2  9.6  7.4 11.0
#> [1079] 15.0 15.0 16.0 15.0 13.0 20.0 16.0 12.0 14.0 21.0 14.0 11.0 10.0  8.0
#> [1093] 18.0 16.0 20.0 17.0 41.0 36.0 87.0 26.0 30.0 16.0 27.0 22.0 19.0 26.0
#> [1107] 26.0 47.0 20.0 55.0 43.0 22.0 35.0 41.0 32.0 22.0 31.0 67.0 17.0 15.0
#> [1121] 15.0 15.0 27.0 42.0 17.0 18.0 18.0 15.0 15.0 17.0 20.0 15.0 15.0 14.0
#> [1135] 11.0 14.0 15.0 11.0  9.3 15.0 15.0 14.0 17.0 30.0 29.0 21.0 14.0 24.0
#> [1149] 18.0 17.0 23.0 18.0 21.0 23.0 23.0 17.0 45.0 14.0 31.0 15.0 24.0 14.0
#> [1163] 33.0 39.0 51.0 20.0 27.0 22.0 20.0 32.0 19.0 20.0 23.0 35.0 35.0 21.0
#> [1177] 36.0 35.0 24.0 23.0 35.0 79.0 56.0 50.0 51.0 75.0 82.0 50.0 34.0 20.0
#> [1191] 22.0 21.0 32.0 30.0 20.0 28.0 36.0 27.0 46.0 22.0 24.0 19.0 50.0 46.0
#> [1205] 42.0 24.0 39.0 25.0 28.0 27.0 24.0 25.0 45.0 37.0 39.0 42.0 33.0 39.0
#> [1219] 64.0 36.0 28.0 23.0 59.0 38.0 27.0 54.0 21.0 26.0 28.0 22.0 19.0 27.0
#> [1233] 38.0 21.0 21.0 37.0 25.0 19.0 22.0 25.0 32.0  8.9 17.0 20.0 22.0 18.0
#> [1247] 20.0 64.0 27.0 29.0 28.0 27.0 28.0 41.0 26.0 13.0 10.0  9.5  7.4 26.0
#> [1261] 15.0 14.0  6.9 14.0 11.0 10.0 11.0 12.0 12.0 11.0 12.0 15.0 18.0 18.0
#> [1275] 20.0 15.0 19.0 17.0 18.0 16.0 19.0 20.0 18.0 22.0 18.0 16.0 15.0 11.0
#> [1289] 16.0 15.0 15.0 19.0 17.0 20.0 22.0 23.0 21.0 13.0 16.0 14.0 19.0 15.0
#> [1303] 12.0 15.0  9.4 12.0 44.0 16.0 12.0  9.9 21.0 14.0 21.0 13.0 56.0 69.0
#> [1317] 82.0 82.0 68.0 74.0 62.0 48.0 46.0 59.0 58.0 87.0 62.0 62.0 22.0 60.0
#> [1331] 34.0 56.0 34.0 76.0 35.0 24.0 38.0 38.0 43.0 38.0 64.0 76.0 56.0 65.0
#> [1345] 76.0 74.0 80.0 86.0 55.0 53.0 73.0 60.0 49.0 52.0 46.0 39.0 36.0 29.0
#> [1359] 13.0 13.0 13.0 16.0 14.0 16.0 14.0 15.0 15.0 21.0 15.0 13.0 14.0  9.0
#> [1373]  6.0  7.0 16.0 15.0  7.0 23.0 12.0 16.0  9.0 10.0 24.0 13.0  7.0  6.0
#> [1387] 17.0 13.0  9.0  7.0  7.0 87.0 71.0 69.0 16.0 52.0 53.0 55.0 71.0 44.0
#> [1401] 59.0 47.0 55.0 38.0 30.0 31.0 32.0 35.0 37.0 32.0 27.0 36.0 39.0 56.0
#> [1415] 39.0 40.0 37.0 42.0 30.0 37.0 41.0 37.0 33.0 38.0 37.0 30.0 37.0 28.0
#> [1429] 34.0 31.0 34.0 31.0 36.0 34.0 39.0 44.0 40.0 45.0 38.0 49.0 38.0 47.0
#> [1443] 52.0 49.0 16.0 18.0 12.0 15.0 13.0 15.0 15.0  8.0  9.0 16.0 10.0 25.0
#> [1457]  6.0 10.0 12.0 14.0  9.0  8.0  8.0 13.0 10.0 43.0 40.0 80.0 66.0 61.0
#> [1471] 73.0 33.0 56.0 51.0 62.0 45.0 47.0 69.0 57.0  9.0 12.0 15.0  8.0  9.0
#> [1485]  9.0  7.0  9.0  8.0  8.0 25.0  6.0 11.0  8.0 14.0 10.0  6.0  8.0 12.0
#> [1499] 10.0  8.0 34.0 36.0 52.0 42.0 10.0  9.0 14.0  3.0  9.0  9.0  7.0  8.0
#> [1513] 10.0  9.0 20.0  9.0 10.0 10.0 18.0 10.0  6.0  8.0  8.0  9.0 41.0 41.0
#> [1527] 43.0 35.0 35.0 50.0 44.0 38.0 44.0 39.0 49.0 47.0 44.0 49.0 49.0 43.0
#> [1541] 74.0 69.0 41.0 40.0 40.0 39.0 35.0 34.0 38.0 42.0 42.0 47.0 36.0 31.0
#> [1555] 29.0 32.0 26.0 29.0 64.0 46.0 40.0 39.0 39.0 52.0 30.0 25.0 30.0 34.0
#> [1569] 52.0 51.0 55.0 51.0 61.0 59.0 43.0 50.0 42.0 51.0 56.0  8.0 10.0 11.0
#> [1583] 14.0 16.0 11.0  8.0 11.0  8.0  9.0 10.0 12.0 32.0  8.0 12.0  8.0 10.0
#> [1597] 11.0  8.0  6.0 10.0  8.0  8.0 59.0 54.0 51.0 54.0 57.0 55.0 49.0 55.0
#> [1611] 53.0 50.0 45.0 14.0 15.0 13.0 14.0 12.0 11.0 13.0 11.0  9.0 10.0  6.0
#> [1625]  7.0 18.0  9.0 23.0  6.0  5.0  4.0  6.0  5.0  8.0  7.0 14.0  8.0  7.0
#> [1639]  6.0  7.0 18.0  5.0  6.0 12.0  5.0 13.0 36.0 41.0 14.0 17.0 14.0 13.0
#> [1653] 14.0 14.0 12.0 14.0 14.0 12.0 84.0 64.0 54.0 54.0 55.0 73.0 39.0 66.0
#> [1667] 81.0 43.0 25.0 48.0 68.0 37.0  6.0 40.0 51.0 29.0 28.0 38.0 24.0 63.0
#> [1681] 50.0 38.0 36.0 38.0 59.0 46.0 59.0 15.0 62.0 86.0 32.0 78.0  1.5 48.0
#> [1695] 62.0 45.0 53.0 73.0 74.0 83.0  1.5 38.0 49.0 25.0 75.0 50.0 28.0 47.0
#> [1709] 30.0 21.0  1.5 60.0 25.0 70.0 89.0 32.0 29.0 40.0 21.0 42.0 57.0 81.0
#> [1723] 41.0 43.0 16.0 33.0 46.0 55.0 52.0 74.0 37.0 46.0 41.0 21.0 37.0 47.0
#> [1737] 16.0 37.0 27.0 26.0 32.0 31.0 21.0 37.0 80.0 61.0 76.0 39.0 40.0 20.0
#> [1751] 33.0 53.0 37.0 40.0 17.0 24.0 54.0 42.0 18.0 27.0 14.0 49.0 33.0 60.0
#> [1765] 21.0 32.0 35.0 40.0 24.0 43.0 81.0 63.0 36.0 15.0  1.5 73.0 46.0 32.0
#> [1779] 38.0 35.0 38.0 23.0 33.0 39.0 70.0 21.0 41.0 33.0 63.0 37.0 46.0 38.0
#> [1793] 33.0 41.0 27.0 20.0 27.0 55.0 52.0 24.0 57.0 20.0 61.0 45.0 73.0 37.0
#> [1807] 73.0 34.0 26.0 35.0 48.0 47.0 10.0 34.0 46.0 29.0 44.0 47.0 65.0 70.0
#> [1821] 38.0 31.0 26.0 48.0 15.0 54.0 29.0 52.0 80.0 24.0 43.0 38.0 37.0 25.0
#> [1835] 18.0 46.0 71.0 73.0 54.0 36.0 37.0 42.0 43.0 30.0 52.0 36.0 42.0 62.0
#> [1849] 65.0 57.0 38.0  1.5 30.0 67.0 42.0 65.0 38.0 30.0 46.0 41.0 29.0 50.0
#> [1863] 46.0 62.0 21.0 72.0 13.0 76.0 28.0 59.0 21.0 27.0 25.0 49.0 43.0 36.0
#> [1877] 63.0 83.0 50.0 58.0 30.0 14.0 61.0 30.0 31.0 10.0 45.0 24.0 39.0 42.0
#> [1891] 54.0 26.0 50.0 39.0 37.0 38.0 57.0 67.0 35.0 35.0 60.0 37.0 46.0 22.0
#> [1905] 63.0 29.0 23.0 29.0 61.0 52.0 36.0 25.0 11.0 34.0 44.0 30.0 54.0 38.0
#> [1919] 68.0 80.0 25.0 31.0 25.0 46.0 48.0 25.0  8.0 32.0 39.0 60.0 27.0 31.0
#> [1933] 43.0 36.0 72.0 46.0 38.0 47.0 50.0 58.0 17.0 57.0 62.0 62.0 32.0 31.0
#> [1947] 48.0 50.0 31.0 50.0 64.0 58.0 28.0 40.0 11.0 39.0 35.0 67.0 35.0 41.0
#> [1961] 41.0 41.0 27.0 14.0 26.0 49.0 14.0 42.0 87.0 24.0 36.0 39.0 24.0 34.0
#> [1975] 46.0 41.0 54.0 52.0 25.0 59.0 33.0 38.0 21.0 41.0 61.0 46.0 21.0 45.0
#> [1989] 45.0 42.0 37.0 58.0 43.0 51.0 51.0 40.0 41.0 54.0 25.0 20.0 24.0 33.0
#> [2003] 49.0 32.0 10.0 54.0 33.0 42.0 10.0 48.0 10.0 33.0 39.0 66.0 68.0 38.0
#> [2017] 26.0 66.0 28.0 70.0 32.0 48.0 47.0 21.0 32.0 41.0 26.0 23.0 48.0 44.0
#> [2031] 56.0 61.0 45.0 36.0 51.0 26.0 39.0 35.0 30.0 45.0 62.0 11.0 43.0 38.0
#> [2045] 24.0 42.0 26.0 41.0 25.0 56.0 49.0 49.0 59.0  1.5 39.0 38.0 50.0 55.0
#> [2059] 55.0 62.0 28.0 34.0 29.0 26.0 46.0 53.0 42.0 32.0 43.0 28.0 39.0 33.0
#> [2073] 25.0 58.0 55.0 53.0 35.0 52.0 61.0 22.0 34.0 30.0 36.0 41.0 50.0 23.0
#> [2087] 67.0 42.0 59.0 19.0 28.0 23.0 41.0 26.0 40.0 47.0 37.0 26.0 70.0 40.0
#> [2101] 38.0 32.0 42.0 56.0 58.0 76.0 36.0 40.0 35.0  9.0 24.0 35.0 38.0 31.0
#> [2115] 45.0 53.0 60.0 22.0 23.0 37.0 48.0 57.0 40.0 50.0 84.0 14.0 46.0 32.0
#> [2129] 73.0 64.0  1.5 28.0 40.0 64.0 33.0 56.0 25.0 46.0 41.0 29.0 50.0 77.0
#> [2143] 69.0  8.0 43.0 71.0 28.0 39.0 23.0 76.0 36.0 37.0 37.0 39.0 17.0 22.0
#> [2157] 76.0 34.0 37.0 25.0 64.0 40.0 37.0 45.0 83.0 26.0 26.0 44.0 42.0 54.0
#> [2171]  6.0 53.0 47.0 35.0  1.5 48.0 30.0 32.0 56.0 44.0 35.0 56.0 83.0 23.0
#> [2185] 38.0 32.0 23.0 40.0 21.0 65.0 24.0 65.0 46.0 71.0 33.0 43.0 54.0 49.0
#> [2199] 64.0 60.0 31.0 24.0 30.0 60.0 41.0 31.0 23.0 58.0 25.0 24.0 30.0 22.0
#> [2213] 18.0 36.0 26.0 22.0 15.0 38.0 37.0 32.0 45.0 16.0  8.0 36.0 33.0 71.0
#> [2227] 76.0 40.0 32.0 64.0 61.0 44.0 43.0 22.0 75.0 55.0 25.0 65.0 35.0 42.0
#> [2241] 50.0 35.0 52.0  1.5 19.0 33.0 17.0  1.5  8.0 57.0  1.5 63.0 30.0 45.0
#> [2255]  1.5 55.0 33.0 65.0 55.0 21.0 18.0 85.0 39.0 57.0 56.0 36.0 32.0 37.0
#> [2269] 46.0 39.0 66.0 72.0 34.0 41.0 41.0 32.0 62.0 24.0 21.0 33.0 44.0 36.0
#> [2283] 51.0 34.0 70.0 34.0 26.0 31.0 27.0 42.0 31.0 33.0 41.0 48.0 22.0 71.0
#> [2297] 66.0 47.0 60.0 48.0 33.0 51.0 15.0 65.0 63.0 32.0 26.0 21.0 48.0 47.0
#> [2311] 31.0  9.0 65.0 38.0 42.0 57.0 40.0 21.0 25.0 54.0 41.0 15.0 58.0 62.0
#> [2325] 35.0 12.0 43.0 28.0 83.0 41.0 42.0 62.0 38.0 58.0 33.0 51.0 46.0 49.0
#> [2339] 24.0 67.0 39.0 42.0 36.0 80.0  1.5 24.0 23.0 22.0 33.0 17.0 36.0 13.0
#> [2353] 45.0 60.0 62.0 57.0 15.0 49.0 23.0 61.0 81.0 31.0 30.0 36.0 38.0 35.0
#> [2367] 43.0 42.0 35.0 48.0 34.0 33.0 41.0 23.0 34.0 86.0 32.0 61.0 22.0 86.0
#> [2381] 55.0 60.0 46.0 64.0 37.0 22.0 46.0 22.0 35.0 39.0 42.0 40.0 89.0 52.0
#> [2395] 50.0 54.0 30.0 38.0 28.0 43.0 14.0 42.0 49.0 53.0 67.0 59.0 64.0
#> 
#> $type
#> [1] "histogram"
#> 
#> $xbins
#> $xbins$start
#> [1] 1.5
#> 
#> 
#> $marker
#> $marker$color
#> [1] "#00bde3"
#> 
#> 
#> $stroke
#> [1] "#005ea2"
#> 
#> $bingroup
#> [1] 1
#> 
#> $name
#> [1] "<b>Outliers Removed</b>\nUpper Threshold: 89\nLower Threshold: 1.5"
#> 
#> $visible
#> [1] "legendonly"
#> 
#> $inherit
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "plotly_histogram" "list"            
#> 
#> 
#> $layout
#> $layout$width
#> NULL
#> 
#> $layout$height
#> NULL
#> 
#> $layout$margin
#> $layout$margin$b
#> [1] 40
#> 
#> $layout$margin$l
#> [1] 60
#> 
#> $layout$margin$t
#> [1] 25
#> 
#> $layout$margin$r
#> [1] 10
#> 
#> 
#> 
#> $source
#> [1] "A"
#> 
#> $config
#> $config$modeBarButtonsToAdd
#> [1] "hoverclosest" "hovercompare"
#> 
#> $config$showSendToCloud
#> [1] FALSE
#> 
#> $config$displayModeBar
#> [1] TRUE
#> 
#> 
#> $layoutAttrs
#> $layoutAttrs$`1bb46ff76a0d`
#> $layoutAttrs$`1bb46ff76a0d`$xaxis
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$title
#> [1] "UG/L"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$titlefont
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$titlefont$size
#> [1] 16
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$titlefont$family
#> [1] "Arial"
#> 
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$tickfont
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$tickfont$size
#> [1] 16
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$tickfont$family
#> [1] "Arial"
#> 
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$hoverformat
#> [1] ",.4r"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$linecolor
#> [1] "black"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$rangemode
#> [1] "tozero"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$showgrid
#> [1] FALSE
#> 
#> $layoutAttrs$`1bb46ff76a0d`$xaxis$tickcolor
#> [1] "black"
#> 
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$title
#> [1] "Frequency"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$titlefont
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$titlefont$size
#> [1] 16
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$titlefont$family
#> [1] "Arial"
#> 
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$tickfont
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$tickfont$size
#> [1] 16
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$tickfont$family
#> [1] "Arial"
#> 
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$hoverformat
#> [1] ",.4r"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$linecolor
#> [1] "black"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$rangemode
#> [1] "tozero"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$showgrid
#> [1] FALSE
#> 
#> $layoutAttrs$`1bb46ff76a0d`$yaxis$tickcolor
#> [1] "black"
#> 
#> 
#> $layoutAttrs$`1bb46ff76a0d`$hoverlabel
#> $layoutAttrs$`1bb46ff76a0d`$hoverlabel$bgcolor
#> [1] "white"
#> 
#> 
#> $layoutAttrs$`1bb46ff76a0d`$title
#> [1] "Total Phosphorus_as P_ug/L"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$plot_bgcolor
#> [1] "#e5ecf6"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$barmode
#> [1] "overlay"
#> 
#> $layoutAttrs$`1bb46ff76a0d`$legend
#> $layoutAttrs$`1bb46ff76a0d`$legend$title
#> $layoutAttrs$`1bb46ff76a0d`$legend$title$text
#> [1] "<b>Select 'Outliers Removed' \nand Deselect 'All Data' \nto View a Subset of the Data<b>"
#> 
#> 
#> 
#> 
#> 
#> attr(,"TOJSON_FUNC")
#> function (x, ...) 
#> {
#>     jsonlite::toJSON(x, digits = 50, auto_unbox = TRUE, force = TRUE, 
#>         null = "null", na = "null", time_format = "%Y-%m-%d %H:%M:%OS6", 
#>         ...)
#> }
#> <bytecode: 0x555ccafbf830>
#> <environment: namespace:plotly>

# Create histogram with additional grouping columns and view first plot in list. In this example, we will group by both the comparable data identifier and the monitoring location type (e.g. stream, reservoir, canal, etc.)
# Load example dataset:
data(Data_Nutrients_Utah)
#> Warning: data set ‘Data_Nutrients_Utah’ not found

Histogram_output = TADA_Histogram(Data_Nutrients_UT, id_cols = c("TADA.ComparableDataIdentifier","MonitoringLocationTypeName"))
#> [1] "Plotting function removed 17 results where TADA.ResultMeasureValue = NA. These results cannot be plotted."
Histogram_output[[1]]