Create Scatterplot(s)

Create Scatterplot(s)

Usage

TADA_Scatterplot(.data, id_cols = c("TADA.ComparableDataIdentifier"))

Arguments

.data

TADA data frame containing the data downloaded from the WQP, where each row represents a unique data record. Data frame must include the columns 'TADA.ComparableDataIdentifier', 'TADA.ResultMeasureValue', and 'TADA.ResultMeasure.MeasureUnitCode' to run this function. The 'TADA.ComparableDataIdentifier' column is added to the data frame automatically when WQP data is retrieved using TADADataRetrieval. This TADA.ComparableDataIdentifier can be updated to harmonize synonyms by running the function TADA_HarmonizeSynonyms(). You can also include additional grouping columns in the id_col input if desired. If more than one TADA.ComparableDataIdentifier exists in the dataset, the function creates a list of plots, where each list element name is a unique TADA.ComparableDataIdentifier.

id_cols

The column(S) in the dataset used to identify the unique groups to be plotted. Defaults to 'TADA.ComparableDataIdentifier'.

Value

A list of plotly scatterplot figures showing the distribution of sample values for each comparable data group (TADA.ComparableDataIdentifier).

Examples

# Load example dataset:
data(Data_6Tribes_5y_Harmonized)

# Create a scatterplot for each comparable data group (TADA.ComparableDataIdentifier)
# in the input dataframe:
TADA_Scatterplot(Data_6Tribes_5y_Harmonized, id_cols = "TADA.ComparableDataIdentifier")
#> $`ALKALINITY, TOTAL_NA_NA_UG/L`
#> Warning: Ignoring 10 observations
#> 
#> $`ALKALINITY, TOTAL_TOTAL_AS CACO3_PCU`
#> Warning: Ignoring 1 observations
#> 
#> $`ALKALINITY, TOTAL_TOTAL_AS CACO3_UG/L`
#> Warning: Ignoring 297 observations
#> 
#> $`ALKALINITY, TOTAL_TOTAL_NA_UG/L`
#> 
#> $`ALPHA PARTICLE_DISSOLVED_NA_PCI/L`
#> 
#> $`ALPHA PARTICLE_NA_NA_PCI/L`
#> 
#> $`ALPHA PARTICLE_TOTAL_NA_PCI/L`
#> 
#> $`ALUMINUM_DISSOLVED_NA_UG/L`
#> 
#> $`AMMONIA_DISSOLVED_AS N_MG/L`
#> 
#> $`AMMONIA_UNFILTERED_AS N_MG/L`
#> Warning: Ignoring 372 observations
#> 
#> $`AMMONIA_UNFILTERED_NA_MG/L`
#> Warning: Ignoring 1 observations
#> 
#> $`AMMONIUM_TOTAL_NA_MG/L`
#> 
#> $`APPARENT COLOR_TOTAL_NA_PCU`
#> Warning: Ignoring 159 observations
#> 
#> $`ARSENIC_DISSOLVED_NA_UG/L`
#> 
#> $`ARSENIC_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 
#> $`BARIUM_DISSOLVED_NA_UG/L`
#> 
#> $`BAROMETRIC PRESSURE_NA_NA_G/M2`
#> 
#> $`BENZENE HEXACHLORIDE, ALPHA (BHC)_TOTAL_NA_UG/L`
#> 
#> $`BERYLLIUM_DISSOLVED_NA_UG/L`
#> 
#> $`BHC, .BETA.-BHC & .GAMMA.-BHC MIX, UNSPECIFIED_TOTAL_NA_UG/L`
#> 
#> $`BORON_DISSOLVED_NA_UG/L`
#> 
#> $`CADMIUM_DISSOLVED_NA_UG/L`
#> 
#> $`CADMIUM_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 
#> $`CALCIUM_DISSOLVED_NA_UG/L`
#> 
#> $`CHEMICAL OXYGEN DEMAND_TOTAL_NA_UG/L`
#> 
#> $`CHLORIDE_TOTAL_NA_UG/L`
#> Warning: Ignoring 112 observations
#> 
#> $`CHLOROPHYLL A, CORRECTED FOR PHEOPHYTIN_SUSPENDED_NA_UG/L`
#> 
#> $`CHLOROPHYLL A_NA_NA_UG/L`
#> Warning: Ignoring 1 observations
#> 
#> $`CHLOROPHYLL A_UNFILTERED_NA_UG/L`
#> Warning: Ignoring 181 observations
#> 
#> $`CHROMIUM_DISSOLVED_NA_UG/L`
#> 
#> $`CHROMIUM_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 
#> $`COBALT_DISSOLVED_NA_UG/L`
#> 
#> $`CONDUCTIVITY_NA_NA_US/CM`
#> 
#> $`COPPER_DISSOLVED_NA_UG/L`
#> 
#> $`COPPER_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 
#> $COUNT_NA_NA_COUNT
#> Warning: Ignoring 1501 observations
#> 
#> $`CYANIDE_DISSOLVED_NA_UG/L`
#> 
#> $`CYANIDE_TOTAL_NA_UG/L`
#> 
#> $`DEPTH, BOTTOM_NA_NA_CM`
#> Warning: Ignoring 403 observations
#> 
#> $`DEPTH, SECCHI DISK DEPTH_NA_NA_FT`
#> 
#> $`DEPTH, SECCHI DISK DEPTH_NA_NA_M`
#> Warning: Ignoring 321 observations
#> 
#> $`DEPTH, SNOW COVER_NA_NA_IN`
#> 
#> $DEPTH_NA_NA_FT
#> 
#> $DEPTH_NA_NA_IN
#> Warning: Ignoring 1 observations
#> 
#> $DEPTH_NA_NA_M
#> Warning: Ignoring 203 observations
#> 
#> $`DISCHARGE, RIVER/STREAM_NA_NA_CFS`
#> 
#> $`DISSOLVED OXYGEN (DO)_NA_NA_MG/L`
#> Warning: Ignoring 450 observations
#> 
#> $`DISSOLVED OXYGEN SATURATION_NA_NA_%`
#> 
#> $`DISSOLVED OXYGEN UPTAKE_NA_NA_UG/L`
#> 
#> $`ENDRIN ALDEHYDE_TOTAL_NA_UG/L`
#> 
#> $`ENDRIN_TOTAL_NA_UG/L`
#> 
#> $`ESCHERICHIA COLI_NA_NA_MPN/100ML`
#> Warning: Ignoring 184 observations
#> 
#> $`FLUORIDE_TOTAL_NA_UG/L`
#> 
#> $`GLYPHOSATE_DISSOLVED_NA_UG/L`
#> 
#> $`GLYPHOSATE_TOTAL_NA_UG/L`
#> 
#> $`HARDNESS, CA, MG_TOTAL_NA_UG/L`
#> Warning: Ignoring 299 observations
#> 
#> $`HARDNESS, CARBONATE_NA_AS CACO3_UG/L`
#> 
#> $`HARDNESS, CARBONATE_NA_NA_UG/L`
#> 
#> $`HARDNESS, CARBONATE_TOTAL_AS CACO3_UG/L`
#> 
#> $`HARDNESS, CARBONATE_TOTAL_NA_UG/L`
#> 
#> $`ICE THICKNESS_NA_NA_IN`
#> 
#> $`IRON_DISSOLVED_NA_UG/L`
#> 
#> $`IRON_TOTAL_NA_UG/L`
#> 
#> $`KJELDAHL NITROGEN_TOTAL RECOVERABLE_NA_UG/L`
#> Warning: Ignoring 1 observations
#> 
#> $`LEAD_DISSOLVED_NA_UG/L`
#> 
#> $`LEAD_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 
#> $`LENGTH, TOTAL (FISH)_NA_NA_MM`
#> Warning: Ignoring 242 observations
#> 
#> $`MAGNESIUM_DISSOLVED_NA_UG/L`
#> 
#> $`MANGANESE_DISSOLVED_NA_UG/L`
#> 
#> $`MANGANESE_TOTAL_NA_UG/L`
#> 
#> $`MERCURY_DISSOLVED_NA_UG/L`
#> Warning: Ignoring 102 observations
#> 
#> $`MERCURY_TOTAL_NA_UG/L`
#> Warning: Ignoring 105 observations
#> 
#> $`MOLYBDENUM_DISSOLVED_NA_UG/L`
#> 
#> $`NICKEL_DISSOLVED_NA_UG/L`
#> 
#> $`NICKEL_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 
#> $`NITRATE + NITRITE_FILTERED_AS N_MG/L`
#> 
#> $`NITRATE + NITRITE_INORGANIC_AS NO3_UG/L`
#> Warning: Ignoring 371 observations
#> 
#> $`NITRATE + NITRITE_UNFILTERED_AS N_MG/L`
#> 
#> $`NITRATE_UNFILTERED_AS N_MG/L`
#> Warning: Ignoring 127 observations
#> 
#> $`NITRATE_UNFILTERED_NA_MG/L`
#> Warning: Ignoring 1 observations
#> 
#> $`NITRITE_UNFILTERED_AS N_MG/L`
#> Warning: Ignoring 112 observations
#> 
#> $`NITRITE_UNFILTERED_NA_MG/L`
#> 
#> $`NON-VIABLE SEED COUNT_NA_NA_COUNT`
#> Warning: Ignoring 65 observations
#> 
#> $`NON-VIABLE SEED WEIGHT_NA_NA_G`
#> Warning: Ignoring 58 observations
#> 
#> $`NUMBER OF PEDICELS PER SAMPLE PLANT_NA_NA_COUNT`
#> Warning: Ignoring 64 observations
#> 
#> $`NUMBER OF SEEDS WITH ERGOTS_NA_NA_COUNT`
#> Warning: Ignoring 9 observations
#> 
#> $`NUMBER OF SEEDS WITH WORM HOLES_NA_NA_COUNT`
#> Warning: Ignoring 27 observations
#> 
#> $`NUMBER OF STALKS PER SAMPLE PLANT_NA_NA_COUNT`
#> Warning: Ignoring 190 observations
#> 
#> $`ORGANIC CARBON_DISSOLVED_NA_UG/L`
#> 
#> $`ORGANIC CARBON_TOTAL_NA_UG/L`
#> 
#> $`ORTHOPHOSPHATE_FILTERED_AS P_UG/L`
#> 
#> $`ORTHOPHOSPHATE_TOTAL RECOVERABLE_AS P_UG/L`
#> Warning: Ignoring 8 observations
#> 
#> $`ORTHOPHOSPHATE_UNFILTERED_AS P_UG/L`
#> Warning: Ignoring 363 observations
#> 
#> $`PERIPHYTON_NA_NA_G/M2`
#> Warning: Ignoring 32 observations
#> 
#> $`PHEOPHYTIN A_TOTAL_NA_UG/L`
#> Warning: Ignoring 182 observations
#> 
#> $`PLANT HEIGHT (ABOVE WATER)_NA_NA_CM`
#> Warning: Ignoring 160 observations
#> 
#> $`PLANT HEIGHT (TOTAL)_NA_NA_CM`
#> Warning: Ignoring 294 observations
#> 
#> $`POTASSIUM_DISSOLVED_NA_UG/L`
#> 
#> $`PRESSURE_NA_NA_G/M2`
#> 
#> $PRESSURE_NA_NA_MMHG
#> Warning: Ignoring 23 observations
#> 
#> $`RADIUM-226_NA_NA_PCI/L`
#> 
#> $`RADIUM-226_TOTAL_NA_PCI/L`
#> 
#> $`RADIUM-228_NA_NA_PCI/L`
#> 
#> $`RADIUM-228_TOTAL_NA_PCI/L`
#> 
#> $`ROOT WEIGHT_NA_NA_G`
#> Warning: Ignoring 65 observations
#> 
#> $SALINITY_NA_NA_PSS
#> 
#> $`SELENIUM_DISSOLVED_NA_UG/L`
#> 
#> $`SELENIUM_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 
#> $`SHOOT WEIGHT_NA_NA_G`
#> Warning: Ignoring 67 observations
#> 
#> $`SILICON_TOTAL_NA_UG/L`
#> 
#> $`SILVER_DISSOLVED_NA_UG/L`
#> 
#> $`SODIUM_DISSOLVED_NA_UG/L`
#> 
#> $`STREAM STAGE_NA_NA_FT`
#> 
#> $`SULFATE_TOTAL_AS SO4_UG/L`
#> 
#> $`SULFATE_TOTAL_NA_UG/L`
#> Warning: Ignoring 169 observations
#> 
#> $`TEMPERATURE, SAMPLE_NA_NA_DEG C`
#> 
#> $`TEMPERATURE, WATER_NA_NA_DEG C`
#> Warning: Ignoring 237 observations
#> 
#> $`THALLIUM_DISSOLVED_NA_UG/L`
#> 
#> $`TOTAL DISSOLVED SOLIDS_NA_NA_UG/KG`
#> 
#> $`TOTAL DISSOLVED SOLIDS_NA_NA_UG/L`
#> Warning: Ignoring 1 observations
#> 
#> $`TOTAL DISSOLVED SOLIDS_TOTAL_NA_UG/L`
#> 
#> $`TOTAL HARDNESS_TOTAL_NA_UG/L`
#> 
#> $`TOTAL KJELDAHL NITROGEN (ORGANIC N & NH3)_UNFILTERED_AS N_MG/L`
#> Warning: Ignoring 372 observations
#> 
#> $`TOTAL NITROGEN, MIXED FORMS_UNFILTERED_AS N_MG/L`
#> 
#> $`TOTAL NITROGEN, MIXED FORMS_UNFILTERED_NA_MG/L`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L`
#> Warning: Ignoring 444 observations
#> 
#> $`TOTAL SAMPLE WEIGHT_NA_NA_G`
#> Warning: Ignoring 121 observations
#> 
#> $`TOTAL SUSPENDED SOLIDS_NA_NA_MG/L`
#> 
#> $`TOTAL SUSPENDED SOLIDS_NON-FILTERABLE (PARTICLE)_NA_MG/L`
#> 
#> $`TOXAPHENE_TOTAL_NA_UG/L`
#> 
#> $`TRANSPARENCY, SECCHI TUBE WITH DISK_NA_NA_CM`
#> 
#> $`TRITIUM_NA_NA_PCI/L`
#> 
#> $TURBIDITY_NA_NA_FNU
#> 
#> $TURBIDITY_NA_NA_FTU
#> Warning: Ignoring 11 observations
#> 
#> $TURBIDITY_UNFILTERED_NA_FTU
#> 
#> $`URANIUM_DISSOLVED_NA_UG/L`
#> 
#> $`VANADIUM_DISSOLVED_NA_UG/L`
#> 
#> $`VIABLE SEED COUNT_NA_NA_COUNT`
#> Warning: Ignoring 66 observations
#> 
#> $`VIABLE SEED WEIGHT_NA_NA_G`
#> Warning: Ignoring 46 observations
#> 
#> $`VOLATILE SUSPENDED SOLIDS_TOTAL_NA_UG/L`
#> 
#> $`ZINC_DISSOLVED_NA_UG/L`
#> 
#> $`ZINC_TOTAL_NA_UG/L`
#> Warning: Ignoring 111 observations
#> 

# Create a single scatterplot using defaults. The input dataframe in this
# example is filtered so it includes only one TADA.ComparableDataIdentifier
df <- dplyr::filter(Data_6Tribes_5y_Harmonized, TADA.ComparableDataIdentifier == "TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L")
TADA_Scatterplot(df, id_cols = "TADA.ComparableDataIdentifier")

# Create multiple scatterplots with additional grouping columns and view the first
# plot in list. In this example, we will group by both TADA.ComparableDataIdentifier
# and MonitoringLocationTypeName (e.g. stream, reservoir, canal, etc.)
# Load example dataset:
data(Data_Nutrients_Utah)
#> Warning: data set ‘Data_Nutrients_Utah’ not found
Scatterplot_output <- TADA_Scatterplot(Data_Nutrients_UT, id_cols = c("TADA.ComparableDataIdentifier", "MonitoringLocationTypeName"))
#> [1] "Plotting function removed 17 results where TADA.ResultMeasureValue = NA. These results cannot be plotted."
# This example generates 45 scatterplots
Scatterplot_output[[10]]

Scatterplot_output[[25]]

Scatterplot_output[[35]]